moved weapons and slimeoid parts out of ewcfg
