+++ ew/static/cfg.py
@@ -4,14 +4,13 @@ from .. import utils as ewutils
 import random
 import json
 import os
-from ..model.weapon import EwWeapon
 from ..model.hunting import EwAttackType
 from ..model.weather import EwWeather
 from ..model.food import EwFood
 from ..model.item import EwItemDef
 from ..model.poi import EwPoi, EwTransportLine
 from ..model.mutation import EwMutationFlavor
-from ..model.slimeoid import EwBody, EwHead, EwMobility, EwOffense, EwDefense, EwSpecial, EwBrain, EwHue
+from ..model.slimeoid import EwHue
 from ..model.quadrants import EwQuadrantFlavor
 from ..model.status import EwStatusEffectDef
 from ..model.farm import EwFarmAction
@@ -3165,257 +3164,6 @@ durability_items = [
 ]
 
 
-def get_weapon_type_stats(weapon_type):
-	types = {
-		"normal": {
-			"damage_multiplier": 1.1,
-			"cost_multiplier": 1,
-			"crit_chance": 0.2,
-			"crit_multiplier": 1.8,
-			"hit_chance": 0.9,
-		},
-		"precision": {
-			"damage_multiplier": 1.3,
-			"cost_multiplier": 1.3,
-			"crit_chance": 0,
-			"crit_multiplier": 2,
-			"hit_chance": 1,
-		},
-		"small_game": {
-			"damage_multiplier": 0.5,
-			"cost_multiplier": 0.25,
-			"crit_chance": 0.1,
-			"crit_multiplier": 2,
-			"hit_chance": 0.95,
-		},
-		"variable_damage": {
-			"damage_multiplier": 0.75,
-			"variable_damage_multiplier": 1.5,
-			"cost_multiplier": 1,
-			"crit_chance": 0.1,
-			"crit_multiplier": 1.5,
-			"hit_chance": 0.9,
-		},
-		"heavy": {
-			"damage_multiplier": 1.7,
-			"cost_multiplier": 2.75,
-			"crit_chance": 0.1,
-			"crit_multiplier": 1.5,
-			"hit_chance": 0.8,
-		},
-		"defensive": {
-			"damage_multiplier": 0.75,
-			"cost_multiplier": 1.5,
-			"crit_chance": 0.1,
-			"crit_multiplier": 1.5,
-			"hit_chance": 0.85,
-		},
-		"burst_fire": {
-			"damage_multiplier": 0.4,
-			"cost_multiplier": 0.8,
-			"crit_chance": 0.2,
-			"crit_multiplier": 1.5,
-			"hit_chance": 0.85,
-			"shots": 3
-		},
-		"minigun": {
-			"damage_multiplier": 0.3,
-			"cost_multiplier": 3,
-			"crit_chance": 0.1,
-			"crit_multiplier": 2,
-			"hit_chance": 0.5,
-			"shots": 10
-		},
-		"incendiary": {
-			"damage_multiplier": 0.75,
-			"bystander_damage": 0.5,
-			"cost_multiplier": 1.5,
-			"crit_chance": 0.1,
-			"crit_multiplier": 2,
-			"hit_chance": 0.9,
-		},
-		"explosive": {
-			"damage_multiplier": 0.5,
-			"bystander_damage": 0.5,
-			"cost_multiplier": 1,
-			"crit_chance": 0.1,
-			"crit_multiplier": 2,
-			"hit_chance": 0.9,
-		},
-		"tool": {
-			"damage_multiplier": 0.5,
-			"cost_multiplier": 1,
-			"crit_chance": 0.2,
-			"crit_multiplier": 1.8,
-			"hit_chance": 0.9,
-		},
-	}
-
-	return types[weapon_type]
-
-def get_normal_attack(weapon_type = "normal", cost_multiplier = None):
-	weapon_stats = get_weapon_type_stats(weapon_type)
-	if cost_multiplier:
-		weapon_stats["cost_multiplier"] = cost_multiplier
-
-	def get_hit_damage(ctn):
-		hit_damage = 0
-		base_damage = ctn.slimes_damage
-
-		player_has_sharptoother = (mutation_id_sharptoother in ctn.user_data.get_mutations())
-		hit_roll = min(random.random(), random.random()) if player_has_sharptoother else random.random()
-		guarantee_crit = (weapon_type == "precision" and ctn.user_data.sidearm == -1)
-
-		if hit_roll < (weapon_stats["hit_chance"] + ctn.hit_chance_mod):
-			effective_multiplier = weapon_stats["damage_multiplier"] 
-			if "variable_damage_multiplier" in weapon_stats:
-				effective_multiplier += random.random() * weapon_stats["variable_damage_multiplier"]
-
-			hit_damage = base_damage * effective_multiplier
-			if guarantee_crit or random.random() < (weapon_stats["crit_chance"] + ctn.crit_mod):
-				hit_damage *= weapon_stats["crit_multiplier"]
-				if not ("shots" in weapon_stats):
-					ctn.crit = True
-		
-		return hit_damage
-
-	def attack(ctn):
-		ctn.slimes_spent = int(ctn.slimes_spent * weapon_stats["cost_multiplier"])
-		damage = 0
-		if "shots" in weapon_stats:
-			ctn.crit = True
-			for _ in range(weapon_stats["shots"]):
-				hit_damage = get_hit_damage(ctn)
-				damage += hit_damage
-				if hit_damage == 0:
-					ctn.crit = False
-		else:
-			damage = get_hit_damage(ctn)
-			if "bystander_damage" in weapon_stats:
-				ctn.bystander_damage = damage * weapon_stats["bystander_damage"]
-		
-		if damage:
-			ctn.slimes_damage = int(damage)
-		else:
-			ctn.miss = True
-
-	return attack
-
-# weapon effect function for "garrote"
-def wef_garrote(ctn = None):
-	ctn.slimes_damage *= 15
-	#ctn.sap_damage = 0
-	#ctn.sap_ignored = ctn.shootee_data.hardened_sap
-
-	user_mutations = ctn.user_data.get_mutations()
-	aim = (random.randrange(100) + 1)
-	if aim <= int(100 * ctn.hit_chance_mod):
-		if mutation_id_sharptoother in user_mutations:
-			if random.random() < 0.5:
-				ctn.miss = True
-		else:
-			ctn.miss = True
-
-	elif aim <= (1 - int(100 * ctn.crit_mod)):
-		ctn.slimes_damage *= 10
-		ctn.crit = True
-
-	if ctn.miss == False:
-		#Stop movement
-		ewutils.moves_active[ctn.user_data.id_user] = 0
-		#Stun player for 5 seconds
-		ctn.user_data.applyStatus(id_status=status_stunned_id, value=(int(ctn.time_now) + 5))
-		#Start strangling target
-		ctn.shootee_data.applyStatus(id_status=status_strangled_id, source=ctn.user_data.id_user)
-
-# weapon effect function for "Eldritch Staff"
-def wef_staff(ctn = None):
-	market_data = ctn.market_data
-	conditions_met = 0
-	conditions = {
-		lambda _: 3 <= market_data.clock < 4, # witching hour
-		lambda _: weather_map.get(market_data.weather) == weather_foggy,
-		lambda _: (market_data.day % 31 == 15 and market_data.clock >= 20) or (market_data.day % 31 == 16 and market_data.clock <= 6), # moonless night
-		lambda ctn: not ctn.user_data.has_soul,
-		lambda ctn: ctn.user_data.get_possession('weapon'),
-		lambda ctn: ctn.user_data.poi == poi_id_thevoid,
-		lambda ctn: ctn.shootee_data.slimes > ctn.user_data.slimes,
-		lambda ctn: (ctn.user_data.salary_credits <= -50000) or (ctn.shootee_data.salary_credits == 0),
-		lambda ctn: (ctn.user_data.poi_death == ctn.user_data.poi) or (ctn.shootee_data.poi_death == ctn.shootee_data.poi),
-		lambda ctn: (ctn.user_data.id_killer == ctn.shootee_data.id_user) or (ctn.user_data.id_user == ctn.shootee_data.id_killer),
-		lambda ctn: (ctn.shootee_data.life_state == life_state_juvenile) or (ctn.shootee_data.life_state == life_state_enlisted and ctn.shootee_data.faction == ctn.user_data.faction),
-	}
-	for condition in conditions:
-		try:
-			if condition(ctn):
-				conditions_met += 1
-		except:
-			pass
-	
-	ctn.slimes_spent = int(ctn.slimes_spent * 2)
-	ctn.slimes_damage = int(ctn.slimes_damage * (0.3 + conditions_met * 0.6))
-	if conditions_met >= (random.randrange(15) + 1): # 6.66% per condition met
-		ctn.crit = True
-		ctn.slimes_damage = int(ctn.slimes_damage * 1.8)
-
-
-def wef_paintgun(ctn = None):
-	ctn.slimes_damage = int(ctn.slimes_damage * .7)
-	ctn.slimes_spent = int(ctn.slimes_spent * .75)
-	aim = (random.randrange(10) + 1)
-	#ctn.sap_ignored = 10
-	#ctn.sap_damage = 2
-
-	if aim >= (9 - int(10 * ctn.crit_mod)):
-		ctn.crit = True
-		ctn.slimes_damage *= 2
-
-def wef_paintroller(ctn = None):
-	ctn.slimes_damage = int(ctn.slimes_damage * 1.75)
-	ctn.slimes_spent = int(ctn.slimes_spent * 4)
-
-	aim = (random.randrange(10) + 1)
-	user_mutations = ctn.user_data.get_mutations()
-
-	if aim <= (1 + int(10 * ctn.hit_chance_mod)):
-		if mutation_id_sharptoother in user_mutations:
-			if random.random() < 0.5:
-				ctn.miss = True
-		else:
-			ctn.miss = True
-
-	elif aim >= (10 - int(10 * ctn.crit_mod)):
-		ctn.crit = True
-		ctn.slimes_damage *= 2
-		#ctn.sap_damage *= 2
-
-def wef_watercolors(ctn = None):
-	ctn.slimes_damage = 4000
-	aim = (random.randrange(250) + 1)
-	user_mutations = ctn.user_data.get_mutations()
-	#ctn.sap_damage = 0
-
-	if aim <= (1 + int(250 * ctn.hit_chance_mod)):
-		if mutation_id_sharptoother in user_mutations:
-			if random.random() < 0.5:
-				ctn.miss = True
-		else:
-			ctn.miss = True
-
-	elif aim == 1000:
-		ctn.crit = True
-		ctn.slimes_damage *= 1
-
-def wef_fingernails(ctn = None):
-	ctn.slimes_damage = int(ctn.slimes_damage * 0.8)
-	aim = (random.randrange(10) + 1)
-	user_mutations = ctn.user_data.get_mutations()
-	#ctn.sap_damage = 2
-	ctn.miss = False
-
-	if aim >= (10 - int(10 * ctn.crit_mod)):
-		ctn.crit = True
-		ctn.slimes_damage *= 2
 
 
 vendor_dojo = "Dojo"
@@ -3430,1170 +3178,6 @@ weapon_class_paint = "paint"
 weapon_class_juvie = "juvie"
 weapon_class_farming = "farming"
 
-weapon_type_convert = {
-weapon_id_watercolors:wef_watercolors,
-weapon_id_spraycan:get_normal_attack(),
-weapon_id_paintroller:wef_paintroller,
-weapon_id_thinnerbomb:get_normal_attack(weapon_type = 'incendiary'),
-weapon_id_paintgun:wef_paintgun,
-weapon_id_paintbrush:get_normal_attack(weapon_type = 'small_game'),
-weapon_id_roomba:get_normal_attack()
-}
-
-# All weapons in the game.
-weapon_list = [
-	EwWeapon( # 1
-		id_weapon = weapon_id_revolver,
-		alias = [
-			"pistol",
-			"handgun",
-			"bigiron"
-		],
-		str_crit = "**Critical Hit!** You have fataly wounded {name_target} with a lethal shot!",
-		str_miss = "**You missed!** Your shot whizzed past {name_target}'s head!",
-		str_equip = "You equip the revolver.",
-		str_name = "revolver",
-		str_weapon = "a revolver",
-		str_weaponmaster_self = "You are a rank {rank} master of the revolver.",
-		str_weaponmaster = "They are a rank {rank} master of the revolver.",
-		#str_trauma_self = "You have scarring on both temples, which occasionally bleeds.",
-		#str_trauma = "They have scarring on both temples, which occasionally bleeds.",
-		str_kill = "{name_player} puts their revolver to {name_target}'s head. **BANG**. Execution-style. Blood splatters across the hot asphalt. {emote_skull}",
-		str_killdescriptor = "gunned down",
-		str_damage = "{name_target} takes a bullet to the {hitzone}!!",
-		str_duel = "**BANG BANG**. {name_player} and {name_target} practice their quick-draw, bullets whizzing past one another's heads.",
-		str_description = "It's a revolver.",
-		str_reload = "You swing out the revolver’s chamber, knocking out the used shells onto the floor before hastily slamming fresh bullets back into it.",
-		str_reload_warning = "**BANG--** *tk tk...* **SHIT!!** {name_player} just spent the last of the ammo in their revolver’s chamber, it’s out of bullets!!",
-		str_scalp = " It has a bullet hole in it.",
-		fn_effect = get_normal_attack(cost_multiplier = 0.8),
-		price = 10000,
-		clip_size = 6,
-		vendors = [vendor_dojo, vendor_breakroom],
-		classes = [weapon_class_ammo],
-		stat = stat_revolver_kills,
-	),
-	EwWeapon( # 2
-		id_weapon = weapon_id_dualpistols,
-		alias = [
-			"dual",
-			"pistols",
-			"berettas",
-			"dualies"
-		],
-		str_crit = "**Critical Hit!** {name_player} has lodged several bullets into {name_target}'s vital arteries!",
-		str_miss = "**You missed!** Your numerous, haphazard shots hit everything but {name_target}!",
-		str_equip = "You equip the dual pistols.",
-		str_name = "dual pistols",
-		str_weapon = "dual pistols",
-		str_weaponmaster_self = "You are a rank {rank} master of the dual pistols.",
-		str_weaponmaster = "They are a rank {rank} master of the dual pistols.",
-		#str_trauma_self = "You have several stitches embroidered into your chest over your numerous bullet wounds.",
-		#str_trauma = "They have several stitches embroidered into your chest over your numerous bullet wounds.",
-		str_kill = "{name_player} dramatically pulls both triggers on their dual pistols midair, sending two bullets straight into {name_target}'s lungs'. {emote_skull}",
-		str_killdescriptor = "double gunned down",
-		str_damage = "{name_target} takes a flurry of bullets to the {hitzone}!!",
-		str_duel = "**tk tk tk tk tk tk tk tk tk tk**. {name_player} and {name_target} hone their twitch aim and trigger fingers, unloading clip after clip of airsoft BBs into one another with the eagerness of small children.",
-		str_description = "They're dual pistols.",
-		str_reload = "You swing out the handles on both of your pistols, knocking out the used magazines onto the floor before hastily slamming fresh mags back into them.",
-		str_reload_warning = "**tk tk tk tk--** *tk...* **SHIT!!** {name_player} just spent the last of the ammo in their dual pistol’s mags, they’re out of bullets!!",
-		str_scalp = " It has a couple bullet holes in it.",
-		fn_effect = get_normal_attack(),
-		price = 10000,
-		vendors = [vendor_dojo, vendor_breakroom],
-		stat = stat_dual_pistols_kills,
-	),
-	EwWeapon( # 3
-		id_weapon = weapon_id_shotgun,
-		alias = [
-			"boomstick",
-			"remington",
-			"scattergun",
-			"r870"
-		],
-		str_crit = "**Critical Hit!** {name_player} has landed a thick, meaty shot into {name_target}'s chest!",
-		str_miss = "**You missed!** Your pellets inexplicably dodge {name_target}. Fucking random bullet spread, this game will never be competitive.",
-		str_equip = "You equip the shotgun.",
-		str_name = "shotgun",
-		str_weapon = "a shotgun",
-		str_weaponmaster_self = "You are a rank {rank} master of the shotgun.",
-		str_weaponmaster = "They are a rank {rank} master of the shotgun.",
-		#str_trauma_self = "You have a few large, gaping holes in your abdomen. Someone could stick their arm through the biggest one.",
-		#str_trauma = "They have a few large, gaping holes in your abdomen. Someone could stick their arm through the biggest one.",
-		str_kill = "{name_player} blasts their shotgun into {name_target}'s chest at point-blank range, causing guts to explode from their back and coat the surrounding street. chk chk Who's next? {emote_skull}",
-		str_killdescriptor = "pumped full of lead",
-		str_damage = "{name_target} takes a shotgun blast to the {hitzone}!!",
-		str_duel = "**BOOM.** {name_player} and {name_target} stand about five feet away from a wall, pumping it full of lead over and over to study it's bullet spread.",
-		str_description = "It's a shotgun.",
-		str_reload = "You tilt your shotgun and pop shell after shell into its chamber before cocking the forend back. Groovy.",
-		str_reload_warning = "**chk--** *...* **SHIT!!** {name_player}’s shotgun has ejected the last shell in it’s chamber, it’s out of ammo!!",
-		str_scalp = " It has a gaping hole in the center.",
-		fn_effect = get_normal_attack(cost_multiplier = 2.5, weapon_type = 'heavy'),
-		clip_size = 2,
-		price = 10000,
-		vendors = [vendor_dojo, vendor_breakroom],
-		classes = [weapon_class_ammo],
-		stat = stat_shotgun_kills,
-	),
-	EwWeapon( # 4
-		id_weapon = weapon_id_rifle,
-		alias = [
-			"assaultrifle",
-			"machinegun",
-			"mg"
-		],
-		str_crit = "**Critical hit!!** You unload an entire magazine into the target!!",
-		str_miss = "**You missed!** Not one of your bullets connected!!",
-		str_equip = "You equip the assault rifle.",
-		str_name = "assault rifle",
-		str_weapon = "an assault rifle",
-		str_weaponmaster_self = "You are a rank {rank} master of the assault rifle.",
-		str_weaponmaster = "They are a rank {rank} master of the assault rifle.",
-		#str_trauma_self = "Your torso is riddled with scarred-over bulletholes.",
-		#str_trauma = "Their torso is riddled with scarred-over bulletholes.",
-		str_kill = "**RAT-TAT-TAT-TAT-TAT!!** {name_player} rains a hail of bullets directly into {name_target}!! They're officially toast! {emote_skull}",
-		str_killdescriptor = "gunned down",
-		str_damage = "Bullets rake over {name_target}'s {hitzone}!!",
-		str_duel = "**RAT-TAT-TAT-TAT-TAT!!** {name_player} and {name_target} practice shooting at distant targets with quick, controlled bursts.",
-		str_description = "It's a rifle.",
-		str_reload = "You hastily rip the spent magazine out of your assault rifle, before slamming a fresh one back into it.",
-		str_reload_warning = "**RAT-TAT-TAT--** *ttrrr...* **SHIT!!** {name_player}’s rifle just chewed up the last of it’s magazine, it’s out of bullets!!",
-		str_scalp = " It has a shit-load of holes in it.",
-		fn_effect = get_normal_attack(cost_multiplier = 0.7, weapon_type = 'burst_fire'),
-		clip_size = 10,
-		price = 10000,
-		vendors = [vendor_dojo, vendor_breakroom],
-		classes = [weapon_class_ammo],
-		stat = stat_rifle_kills,
-	),
-	EwWeapon( # 5
-		id_weapon = weapon_id_smg,
-		alias = [
-			"submachinegun",
-			"machinegun"
-		],
-		str_crit = "**Critical hit!!** {name_target}’s vital arteries are ruptured by miraculously accurate bullets that actually hit their intended target!!",
-		str_miss = "**You missed!!** {name_player}'s reckless aiming sends their barrage of bullets in every direction but into {name_target}’s body!",
-		str_equip = "You equip the SMG.",
-		str_name = "SMG",
-		str_weapon = "an SMG",
-		str_weaponmaster_self = "You are a rank {rank} master of the SMG.",
-		str_weaponmaster = "They are a rank {rank} master of the SMG.",
-		#str_trauma_self = "Your copious amount of bullet holes trigger onlookers’ Trypophobia.",
-		#str_trauma = "Their copious amount of bullet holes trigger onlookers’ Trypophobia.",
-		str_kill = "**RATTA TATTA TAT!!** {name_player}’s bullet rip through what little was left of {name_target} after the initial barrage. All that remains is a few shreds of clothing and splatterings of slime. {emote_skull}",
-		str_killdescriptor = "riddled with bullets",
-		str_damage = "A reckless barrage of bullets pummel {name_target}’s {hitzone}!!",
-		str_duel = "**RATTA TATTA TAT!!** {name_player} and {name_target} spray bullets across the floor and walls of the Dojo, having a great time.",
-		str_description = "It's a submachine gun.",
-		# str_jammed = "Your SMG jams again, goddamn piece of shit gun...",
-		str_reload = "You hastily rip the spent magazine out of your SMG, before slamming a fresh one back into it.",
-		str_reload_warning = "**RATTA TATTA--** *tk tk tk tk…* **SHIT!!** {name_player}’s SMG just chewed up the last of it’s magazine, it’s out of bullets!!",
-		# str_unjam = "{name_player} successfully whacks their SMG hard enough to dislodge whatever hunk of gunk was blocking it’s internal processes.",
-		str_scalp = " It has a bunch of holes strewn throughout it.",
-		fn_effect = get_normal_attack(cost_multiplier = 0.7, weapon_type = 'burst_fire'),
-		clip_size = 10,
-		price = 10000,
-		vendors = [vendor_dojo, vendor_breakroom],
-		classes = [weapon_class_ammo],
-		stat = stat_smg_kills,
-	),
-	EwWeapon( # 6
-		id_weapon = weapon_id_minigun,
-		alias = [
-			"mini",
-			"gatlinggun"
-		],
-		str_crit = "**Critical hit!!** Round after round of bullets fly through {name_target}, inflicting irreparable damage!!",
-		str_miss = "**You missed!!** Despite the growing heap of used ammunition shells {name_player} has accrued, none of their bullets actually hit {name_target}!",
-		str_equip = "You equip the minigun.",
-		str_name = "minigun",
-		str_weapon = "a minigun",
-		str_weaponmaster_self = "You are a rank {rank} master of the minigun.",
-		str_weaponmaster = "They are a rank {rank} master of the minigun.",
-		#str_trauma_self = "What little is left of your body has large holes punched through it, resembling a slice of swiss cheese.",
-		#str_trauma = "What little is left of their body has large holes punched through it, resembling a slice of swiss cheese.",
-		str_kill = "**TKTKTKTKTKTKTKTKTK!!** {name_player} pushes their minigun barrel right up to {name_target}’s chest, unloading a full round of ammunition and knocking their lifeless corpse back a few yards from the sheer force of the bullets. They failed to outsmart bullet. {emote_skull}",
-		str_killdescriptor = "obliterated",
-		str_damage = "Cascades of bullet easily puncture and rupture {name_target}’s {hitzone}!!",
-		str_duel = "**...** {name_player} and {name_target} crouch close to the ground, throwing sandwiches unto the floor next to each other and repeating memetic voice lines ad nauseam.",
-		str_description = "It's a minigun.",
-		#str_reload = "You curse under your breath, before pulling a fresh belt of bullets from hammerspace and jamming it into your minigun’s hungry feed.",
-		#str_reload_warning = "**TKTKTKTKTKTK--** *wrrrrrr…* **SHIT!!** {name_player}’s minigun just inhaled the last of it’s belt, it’s out of bullets!!",
-		str_scalp = " It looks more like a thick slice of swiss cheese than a scalp.",
-		fn_effect = get_normal_attack(weapon_type = 'minigun'),
-		price = 1000000,
-		vendors = [vendor_bazaar],
-		classes= [weapon_class_captcha],
-		stat = stat_minigun_kills,
-		captcha_length = 6
-	),
-	EwWeapon( # 7
-		id_weapon = weapon_id_bat,
-		alias = [
-			"club",
-			"batwithnails",
-			"nailbat",
-		],
-		str_crit = "**Critical hit!!** {name_player} has bashed {name_target} up real bad!",
-		str_miss = "**MISS!!** {name_player} swung wide and didn't even come close!",
-		str_equip = "You equip the bat with nails in it.",
-		str_name = "bat",
-		str_weaponmaster_self = "You are a rank {rank} master of the nailbat.",
-		str_weaponmaster = "They are a rank {rank} master of the nailbat.",
-		str_weapon = "a bat full of nails",
-		#str_trauma_self = "Your head appears to be slightly concave on one side.",
-		#str_trauma = "Their head appears to be slightly concave on one side.",
-		str_kill = "{name_player} pulls back for a brutal swing! **CRUNCCHHH.** {name_target}'s brains splatter over the sidewalk. {emote_skull}",
-		str_killdescriptor = "nail bat battered",
-		str_damage = "{name_target} is struck with a hard blow to the {hitzone}!!",
-		# str_backfire = "{name_player} recklessly budgens themselves with a particularly overzealous swing! Man, how the hell could they fuck up so badly?",
-		str_duel = "**SMASHH! CRAASH!!** {name_player} and {name_target} run through the neighborhood, breaking windshields, crushing street signs, and generally having a hell of a time.",
-		str_description = "It's a nailbat.",
-		str_scalp = " It has a couple nails in it.",
-		fn_effect = get_normal_attack(weapon_type = 'variable_damage'),
-		price = 10000,
-		vendors = [vendor_dojo, vendor_breakroom],
-		stat = stat_bat_kills,
-	),
-	EwWeapon( # 8
-		id_weapon = weapon_id_brassknuckles,
-		alias = [
-			"knuckles",
-			"knuckledusters",
-			"dusters"
-		],
-		str_crit = "***SKY UPPERCUT!!*** {name_player} executes an artificially difficult combo, rocketing their fist into the bottom of {name_target}’s jaw so hard that {name_target}’s colliding teeth brutally sever an inch off their own tongue!!",
-		str_miss = "**MISS!** {name_player} couldn't land a single blow!!",
-		str_equip = "You equip the brass knuckles.",
-		str_name = "brass knuckles",
-		str_weapon = "brass knuckles",
-		str_weaponmaster_self = "You are a rank {rank} master pugilist.",
-		str_weaponmaster = "They are a rank {rank} master pugilist.",
-		#str_trauma_self = "You've got two black eyes, missing teeth, and a profoundly crooked nose.",
-		#str_trauma = "They've got two black eyes, missing teeth, and a profoundly crooked nose.",
-		str_kill = "{name_player} slugs {name_target} right between the eyes! *POW! THWACK!!* **CRUNCH.** Shit. May have gotten carried away there. Oh, well. {emote_skull}",
-		str_killdescriptor = "pummeled to death",
-		str_damage = "{name_target} is socked in the {hitzone}!!",
-		str_duel = "**POW! BIFF!!** {name_player} and {name_target} take turns punching each other in the abs. It hurts so good.",
-		str_description = "They're brass knuckles.",
-		str_scalp = " It has bone fragments in it.",
-		fn_effect = get_normal_attack(weapon_type = 'variable_damage'),
-		price = 10000,
-		vendors = [vendor_dojo, vendor_breakroom],
-		stat = stat_brassknuckles_kills,
-	),
-	EwWeapon( # 9
-		id_weapon = weapon_id_katana,
-		alias = [
-			"weebsword",
-			"ninjasword",
-			"samuraisword",
-			"blade"
-		],
-		str_crit = "**Critical hit!!** {name_target} is cut deep!!",
-		str_miss = "",
-		str_equip = "You equip the katana.",
-		str_name = "katana",
-		str_weapon = "a katana",
-		str_weaponmaster_self = "You are a rank {rank} blademaster.",
-		str_weaponmaster = "They are a rank {rank} blademaster.",
-		#str_trauma_self = "A single clean scar runs across the entire length of your body.",
-		#str_trauma = "A single clean scar runs across the entire length of their body.",
-		str_kill = "Faster than the eye can follow, {name_player}'s blade glints in the greenish light. {name_target} falls over, now in two pieces. {emote_skull}",
-		str_killdescriptor = "bisected",
-		str_damage = "{name_target} is slashed across the {hitzone}!!",
-		str_duel = "**CRACK!! THWACK!! CRACK!!** {name_player} and {name_target} duel with bamboo swords, viciously striking at head, wrist and belly.",
-		str_description = "It's a katana.",
-		str_scalp = " It seems to have been removed with some precision.",
-		fn_effect = get_normal_attack(weapon_type = 'precision'),
-		price = 10000,
-		vendors = [vendor_dojo, vendor_breakroom],
-		classes= [weapon_class_captcha],
-		stat = stat_katana_kills,
-		captcha_length = 4
-	),
-	EwWeapon( # 10
-		id_weapon = weapon_id_broadsword,
-		alias = [
-			"sword",
-			"highlander",
-			"arawheapofiron",
-			"eyelander"
-		],
-		str_crit = "Critical hit!! {name_player} screams at the top of their lungs and unleashes a devastating overhead swing that maims {name_target}.",
-		str_miss = "You missed! You grunt as your failed overhead swing sends ripples through the air.",
-		# str_backfire = "You feel the bones in your wrists snap as you botch your swing with the heavy blade!! Fucking ouch dawg!",
-		str_equip = "You equip the broadsword.",
-		str_name = "broadsword",
-		str_weapon = "a broadsword",
-		str_weaponmaster_self = "You are a rank {rank} berserker.",
-		str_weaponmaster = "They are a rank {rank} berserker.",
-		#str_trauma_self = "A large dent resembling that of a half-chopped down tree appears on the top of your head.",
-		#str_trauma = "A dent resembling that of a half-chopped down tree appears on the top of their head.",
-		str_kill = "{name_player} skewers {name_target} through the back to the hilt of their broadsword, before kicking their lifeless corpse onto the street corner in gruseome fashion. {name_player} screams at the top of their lungs. {emote_skull}",
-		str_killdescriptor = "slayed",
-		str_damage = "{name_target}'s {hitzone} is separated from their body!!",
-		str_duel = "SCHWNG SCHWNG! {name_player} and {name_target} scream at the top of their lungs to rehearse their battle cries.",
-		str_description = "It's a broadsword.",
-		str_reload = "You summon strength and muster might from every muscle on your body to hoist your broadsword up for another swing.",
-		str_reload_warning = "**THUD...** {name_player}’s broadsword is too heavy, it’s blade has fallen to the ground!!",
-		str_scalp = " It was sloppily lopped off.",
-		clip_size = 1,
-		price = 10000,
-		fn_effect = get_normal_attack(weapon_type = 'heavy'),
-		vendors = [vendor_dojo, vendor_breakroom],
-		stat = stat_broadsword_kills,
-	),
-	EwWeapon( # 11
-		id_weapon = weapon_id_nunchucks,
-		alias = [
-			"nanchacku",
-			"nunchaku",
-			"chucks",
-			"numchucks",
-			"nunchucks"
-		],
-		str_crit = "**COMBO!** {name_player} strikes {name_target} with a flurry of 5 vicious blows!",
-		# str_backfire = "**Whack!!** {name_player} fucks up their kung-fu routine and whacks themselves in the head with their own nun-chucks!!",
-		str_miss = "**WOOSH** {name_player} whiffs every strike!",
-		str_equip = "You equip the nun-chucks.",
-		str_name = "nun-chucks",
-		str_weapon = "nun-chucks",
-		str_weaponmaster_self = "You are a rank {rank} kung-fu master.",
-		str_weaponmaster = "They are a rank {rank} kung-fu master.",
-		#str_trauma_self = "You are covered in deep bruises. You hate martial arts of all kinds.",
-		#str_trauma = "They are covered in deep bruises. They hate martial arts of all kinds.",
-		str_kill = "**HIIII-YAA!!** With expert timing, {name_player} brutally batters {name_target} to death, then strikes a sweet kung-fu pose. {emote_skull}",
-		str_killdescriptor = "fatally bludgeoned",
-		str_damage = "{name_target} takes a bunch of nun-chuck whacks directly in the {hitzone}!!",
-		str_duel = "**HII-YA! HOOOAAAAAHHHH!!** {name_player} and {name_target} twirl wildly around one another, lashing out with kung-fu precision.",
-		str_description = "They're nunchucks.",
-		str_scalp = " It looks very bruised.",
-		fn_effect = get_normal_attack(weapon_type = 'burst_fire'),
-		price = 10000,
-		vendors = [vendor_dojo, vendor_breakroom],
-		stat = stat_nunchucks_kills,
-	),
-	EwWeapon( # 12
-		id_weapon = weapon_id_scythe,
-		alias = [
-			"sickle"
-		],
-		str_crit = "**Critical hit!!** {name_target} is carved by the wicked curved blade!",
-		str_miss = "**MISS!!** {name_player}'s swings wide of the target!",
-		str_equip = "You equip the scythe.",
-		str_name = "scythe",
-		str_weapon = "a scythe",
-		str_weaponmaster_self = "You are a rank {rank} master of the scythe.",
-		str_weaponmaster = "They are a rank {rank} master of the scythe.",
-		#str_trauma_self = "You are wrapped tightly in bandages that hold your two halves together.",
-		#str_trauma = "They are wrapped tightly in bandages that hold their two halves together.",
-		str_kill = "**SLASHH!!** {name_player}'s scythe cleaves the air, and {name_target} staggers. A moment later, {name_target}'s torso topples off their waist. {emote_skull}",
-		str_killdescriptor = "sliced in twain",
-		str_damage = "{name_target} is cleaved through the {hitzone}!!",
-		str_duel = "**WHOOSH, WHOOSH** {name_player} and {name_target} swing their blades in wide arcs, dodging one another's deadly slashes.",
-		str_description = "It's a scythe.",
-		str_scalp = " It's cut in two pieces.",
-		price = 10000,
-		fn_effect = get_normal_attack(weapon_type = 'heavy'),
-		vendors = [vendor_dojo, vendor_breakroom],
-		stat = stat_scythe_kills,
-	),
-	EwWeapon( # 13
-		id_weapon = weapon_id_yoyo,
-		alias = [
-			"yo-yos",
-			"yoyo",
-			"yoyos"
-		],
-		str_crit = "SMAAAASH!! {name_player} pulls off a modified Magic Drop, landing a critical hit on {name_target} just after the rejection!",
-		str_miss = "You missed! {name_player} misjudges their yo-yos trajectory and botches an easy trick.",
-		str_equip = "You equip the yo-yo.",
-		str_name = "yo-yo",
-		str_weaponmaster_self = "You are a rank {rank} master of the yo-yo.",
-		str_weaponmaster = "They are a rank {rank} master of the yo-yo.",
-		str_weapon = "a yo-yo",
-		#str_trauma_self = "Simple yo-yo tricks caught even in your peripheral vision triggers intense PTSD flashbacks.",
-		#str_trauma = "Simple yo-yo tricks caught even in their peripheral vision triggers intense PTSD flashbacks.",
-		str_kill = "{name_player} performs a modified Kwyjibo, effortlessly nailing each step before killing their opponent just ahead of the dismount.",
-		str_killdescriptor = "amazed",
-		str_damage = "{name_player} used {name_target}'s {hitzone} as a counterweight!!",
-		str_duel = "whhzzzzzz {name_player} and {name_target} practice trying to Walk the Dog for hours. It never clicks.",
-		str_description = "It's a yo-yo.",
-		str_scalp = " It has a ball bearing hidden inside it. You can spin it like a fidget spinner.",
-		fn_effect = get_normal_attack(),
-		price = 10000,
-		vendors = [vendor_dojo, vendor_breakroom],
-		stat = stat_yoyo_kills,
-	),
-	EwWeapon( # 14
-		id_weapon = weapon_id_knives,
-		alias = [
-			"knife",
-			"dagger",
-			"daggers",
-			"throwingknives",
-			"throwingknife"
-		],
-		str_crit = "**Critical hit!!** {name_player}'s knife strikes a vital point!",
-		str_miss = "**MISS!!** {name_player}'s knife missed its target!",
-		str_equip = "You equip the throwing knives.",
-		str_name = "throwing knives",
-		str_weapon = "throwing knives",
-		str_weaponmaster_self = "You are a rank {rank} master of the throwing knife.",
-		str_weaponmaster = "They are a rank {rank} master of the throwing knife.",
-		#str_trauma_self = "You are covered in scarred-over lacerations and puncture wounds.",
-		#str_trauma = "They are covered in scarred-over lacerations and puncture wounds.",
-		str_kill = "A blade flashes through the air!! **THUNK.** {name_target} is a goner, but {name_player} slits their throat before fleeing the scene, just to be safe. {emote_skull}",
-		str_killdescriptor = "knifed",
-		str_damage = "{name_target} is stuck by a knife in the {hitzone}!!",
-		str_duel = "**TING! TING!!** {name_player} and {name_target} take turns hitting one another's knives out of the air.",
-		str_description = "They're throwing knives.",
-		str_scalp = " It has about a half dozen stab holes in it.",
-		fn_effect = get_normal_attack(weapon_type = 'small_game'),
-		price = 10000,
-		vendors = [vendor_dojo, vendor_breakroom],
-		stat = stat_knives_kills,
-	),
-	EwWeapon( # 15
-		id_weapon = weapon_id_molotov,
-		alias = [
-			"firebomb",
-			"molotovcocktail",
-			"bomb",
-			"bombs",
-			"moly"
-		],
-		# str_backfire = "**Oh, the humanity!!** The bottle bursts in {name_player}'s hand, burning them terribly!!",
-		str_miss = "**A dud!!** the rag failed to ignite the molotov!",
-		str_crit = "{name_player}’s cocktail shatters at the feet of {name_target}, sending a shower of shattered shards of glass into them!!",
-		str_equip = "You equip the molotov cocktail.",
-		str_name = "molotov cocktail",
-		str_weapon = "molotov cocktails",
-		str_weaponmaster_self = "You are a rank {rank} master arsonist.",
-		str_weaponmaster = "They are a rank {rank} master arsonist.",
-		#str_trauma_self = "You're wrapped in bandages. What skin is showing appears burn-scarred.",
-		#str_trauma = "They're wrapped in bandages. What skin is showing appears burn-scarred.",
-		str_kill = "**SMASH!** {name_target}'s front window shatters and suddenly flames are everywhere!! The next morning, police report that {name_player} is suspected of arson. {emote_skull}",
-		str_killdescriptor = "exploded",
-		str_damage = "{name_target} dodges a bottle, but is singed on the {hitzone} by the blast!!",
-		str_duel = "{name_player} and {name_target} compare notes on frontier chemistry, seeking the optimal combination of combustibility and fuel efficiency.",
-		str_description = "These are glass bottles filled with some good ol' fashioned pyrotechnics.",
-		str_scalp = " It's burnt to a crisp!",
-		fn_effect = get_normal_attack(weapon_type = 'incendiary'),
-		price = 10000,
-		vendors = [vendor_dojo, vendor_breakroom],
-		classes = [weapon_class_burning, weapon_class_captcha],
-		stat = stat_molotov_kills,
-		captcha_length = 4
-	),
-	EwWeapon( # 16
-		id_weapon = weapon_id_grenades,
-		alias = [
-			"nades",
-			"grenade"
-		],
-		str_crit = "**Critical hit!!** {name_target} is blown off their feet by the initial explosion, and lacerated by innumerable shards of shrapnel scattering themselves through their body!!",
-		str_miss = "**You missed!!** {name_player}’s poor aim sends their grenade into a nearby alleyway, it’s explosion eliciting a Wilhelm scream and the assumed death of an innocent passerby. LOL!!",
-		str_equip = "You equip the grenades.",
-		str_name = "grenades",
-		str_weapon = "grenades",
-		str_weaponmaster_self = "You are a rank {rank} master of the grenades.",
-		str_weaponmaster = "They are a rank {rank} master of the grenades.",
-		#str_trauma_self = "Blast scars and burned skin are spread unevenly across your body.",
-		#str_trauma = "Blast scars and burned skin are spread unevenly across their body.",
-		str_kill = "**KA-BOOM!!** {name_player} pulls the safety pin and holds their grenade just long enough to cause it to explode mid air, right in front of {name_target}’s face, blowing it to smithereens. {emote_skull}",
-		str_killdescriptor = "exploded",
-		str_damage = "{name_player}’s grenade explodes, sending {name_target}’s {hitzone} flying off their body!!",
-		str_duel = "**KA-BOOM!!** {name_player} and {name_target} pull the pin out of their grenades and hold it in their hands to get a feel for how long it takes for them to explode. They lose a few body parts in the process.",
-		str_description = "A stack of grenades.",
-		str_scalp = " It's covered in metallic shrapnel.",
-		fn_effect = get_normal_attack(weapon_type = 'explosive'),
-		price = 10000,
-		vendors = [vendor_dojo, vendor_breakroom],
-		classes = [weapon_class_exploding, weapon_class_captcha],
-		stat = stat_grenade_kills,
-		captcha_length = 4
-	),
-	EwWeapon( # 17
-		id_weapon = weapon_id_garrote,
-		alias = [
-			"wire",
-			"garrotewire",
-			"garrottewire"
-		],
-		str_crit = "**CRITICAL HIT!!** {name_player} got lucky and caught {name_target} completely unaware!!",
-		str_miss = "**MISS!** {name_player}'s target got away in time!",
-		str_equip = "You equip the garrotte wire.",
-		str_name = "garrote wire",
-		str_weapon = "a garrotte wire",
-		str_weaponmaster_self = "You are a rank {rank} master of the garrotte.",
-		str_weaponmaster = "They are a rank {rank} master of the garrotte.",
-		#str_trauma_self = "There is noticeable bruising and scarring around your neck.",
-		#str_trauma = "There is noticeable bruising and scarring around their neck.",
-		str_kill = "{name_player} quietly moves behind {name_target} and... **!!!** After a brief struggle, only a cold body remains. {emote_skull}",
-		str_killdescriptor = "garrote wired",
-		str_damage = "{name_target} is ensnared by {name_player}'s wire!!",
-		str_duel = "{name_player} and {name_target} compare their dexterity by playing Cat's Cradle with deadly wire.",
-		str_description = "It's a garrote wire.",
-		str_scalp = " It's a deep shade of blue.",
-		fn_effect = wef_garrote,
-		price = 10000,
-		vendors = [vendor_dojo, vendor_breakroom],
-		stat = stat_garrote_kills,
-	),
-	EwWeapon(  # 18
-		id_weapon = weapon_id_pickaxe,
-		alias = [
-			"pick",
-			"poudrinpickaxe",
-			"poudrinpick"
-		],
-		str_crit = "**Critical hit!!** By sheer dumb luck, {name_player} manages to get a good hit off on {name_target}’s {hitzone}.",
-		str_miss = "**MISS!!** {name_player} is too weak to lift their pickaxe!",
-		str_equip = "You equip the pickaxe.",
-		str_name = "pickaxe",
-		str_weapon = "a pickaxe",
-		str_weaponmaster_self = "You are a rank {rank} coward of the pickaxe.",
-		str_weaponmaster = "They are a rank {rank} coward of the pickaxe.",
-		#str_trauma_self = "There is a deep, precise indent in the crown of your skull. How embarrassing!",
-		#str_trauma = "There is a deep, precise indent in the crown of their skull. How embarrassing!",
-		str_kill = "**THWACK!!** {name_player} summons what little courage they possess to lift the pickaxe above their head and !mine {name_target} to death. How embarrassing! {emote_skull}",
-		str_killdescriptor = "!mined",
-		str_damage = "{name_target} is lightly tapped on the {hitzone}!!",
-		str_duel = "**THWACK, THWACK** {name_player} and {name_target} spend some quality time together, catching up and discussing movies they recently watched or food they recently ate.",
-		str_scalp = " It reeks of dirt and poudrins. How embarrassing!",
-		fn_effect = get_normal_attack(weapon_type = 'tool'),
-		str_description = "It's a pickaxe.",
-		acquisition = acquisition_smelting,
-		stat = stat_pickaxe_kills,
-		is_tool = 1
-	),
-	EwWeapon(  # 19
-		id_weapon = weapon_id_fishingrod,
-		alias = [
-			"fish",
-			"fishing",
-			"rod",
-			"super",
-			"superrod",
-			"superfishingrod"
-		],
-		str_crit = "**Critical hit!!** By sheer dumb luck, {name_player} manages to get a good hit off on {name_target}’s {hitzone}.",
-		str_miss = "**MISS!!** {name_player} is too weak to cast their fishing rod!",
-		str_equip = "You equip the super fishing rod.",
-		str_name = "super fishing rod",
-		str_weapon = "a super fishing rod",
-		str_weaponmaster_self = "You are a rank {rank} coward of the super fishing rod.",
-		str_weaponmaster = "They are a rank {rank} coward of the super fishing rod.",
-		#str_trauma_self = "There is a piercing on the side of your mouth. How embarrassing!",
-		#str_trauma = "There is a piercing on the side of their mouth. How embarrassing!",
-		str_kill = "*whsssh* {name_player} summons what little courage they possess to reel in {name_target} and wring all the slime out of them. How embarrassing! {emote_skull}",
-		str_killdescriptor = "!reeled",
-		str_damage = "{name_target} is lightly pierced on the {hitzone}!!",
-		str_duel = "**whsssh, whsssh** {name_player} and {name_target} spend some quality time together, discussing fishing strategy and preferred types of bait.",
-		str_scalp = " It has a fishing hook stuck in it. How embarrassing!",
-		fn_effect = get_normal_attack(weapon_type = 'tool'),
-		str_description = "It's a super fishing rod.",
-		acquisition = acquisition_smelting,
-		stat = stat_fishingrod_kills,
-		is_tool = 1
-	),
-	EwWeapon(  # 20
-		id_weapon = weapon_id_bass,
-		alias = [
-			"bass",
-		],
-		str_crit = "**Critical hit!!** Through skilled swipes {name_player} manages to sharply strike {name_target}’s {hitzone}.",
-		str_miss = "**MISS!!** {name_player} swings and misses like a dumbass!",
-		str_equip = "You equip the bass guitar, a highly distorted and reverbed riff of unknown origin plays as you place the strap over your neck.",
-		str_name = "bass guitar",
-		str_weapon = "a bass guitar",
-		str_weaponmaster_self = "You are a rank {rank} master of the bass guitar.",
-		str_weaponmaster = "They are a rank {rank} master of the bass guitar.",
-		#str_trauma_self = "There is a large concave dome in the side of your head.",
-		#str_trauma = "There is a large concave dome in the side of their head.",
-		str_kill = "*CRASSHHH.* {name_player} brings down the bass on {name_target} with righteous fury. Discordant notes play harshly as the bass trys its hardest to keep itself together. {emote_skull}",
-		str_killdescriptor = "smashed to pieces",
-		str_damage = "{name_target} is whacked across the {hitzone}!!",
-		str_duel = "**SMASHHH.** {name_player} and {name_target} smash their bass together before admiring eachothers skillful basslines.",
-		str_scalp = " If you listen closely, you can still hear the echoes of a sick bassline from yesteryear.",
-		fn_effect = get_normal_attack(weapon_type = 'variable_damage'),
-		str_description = "It's a bass guitar. All of its strings are completely out of tune and rusted.",
-		acquisition = acquisition_smelting,
-		stat = stat_bass_kills,
-	),
-	EwWeapon(  # 21
-		id_weapon = weapon_id_umbrella,
-		alias = [
-			"umbrella",
-			"slimebrella",
-			"slimecorpumbrella"
-		],
-		str_crit = "**Critical hit!!** {name_player} briefly stuns {name_target} by opening their umbrella in their face, using the opportunity to score a devastating blow to their {hitzone}.",
-		str_miss = "**MISS!!** {name_player} fiddles with their umbrella, failing to open it!",
-		str_equip = "You equip the umbrella.",
-		str_name = "umbrella",
-		str_weapon = "an umbrella",
-		str_weaponmaster_self = "You are a rank {rank} master of the umbrella.",
-		str_weaponmaster = "They are a rank {rank} master of the umbrella.",
-		#str_trauma_self = "You have a large hole in your chest.",
-		#str_trauma = "They have a large hole in their chest.",
-		str_kill = "*SPLAT.* {name_player} pierces {name_target} through the chest, hoists them over their head and opens their umbrella, causing them to explode in a rain of blood and slime. {emote_skull}",
-		str_killdescriptor = "umbrella'd",
-		str_damage = "{name_target} is struck in the {hitzone}!!",
-		str_duel = "**THWACK THWACK.** {name_player} and {name_target} practice their fencing technique, before comparing their favorite umbrella patterns.",
-		str_scalp = " At least it didn't get wet.",
-		fn_effect = get_normal_attack(weapon_type = 'defensive'),
-		str_description = "It's an umbrella, both stylish and deadly.",
-		price = 100000,
-		vendors = [vendor_bazaar],
-		classes = [weapon_class_defensive, weapon_class_captcha],
-		stat = stat_umbrella_kills,
-		captcha_length = 4
-	),
-	EwWeapon(  # 22
-		id_weapon = weapon_id_bow,
-		alias = [
-			"bow",
-		],
-		str_crit = "**Critical hit!!** Through measured shots {name_player} manages to stick a pixelated arrow in {name_target}’s {hitzone}.",
-		str_miss = "**MISS!!** {name_player} completely misses, a pixelated arrow embeds itself into the ground!",
-		str_equip = "You equip the minecraft bow, c418 music plays in the background.",
-		str_name = "minecraft bow",
-		str_weapon = "a minecraft bow",
-		str_weaponmaster_self = "You are a rank {rank} minecraft bowmaster.",
-		str_weaponmaster = "They are a rank {rank} minecraft bowmaster.",
-		#str_trauma_self = "There is a pixelated arrow in the side of your head.",
-		#str_trauma = "There is a pixelated arrow in the side of their head.",
-		str_kill = "*Pew Pew Pew.* {name_player} spams the bow as {name_target}'s life fades, riddling their body with arrows. {emote_skull}",
-		str_killdescriptor = "shot to death",
-		str_damage = "{name_target} is shot in the {hitzone}!!",
-		str_duel = "{name_player} and {name_target} shoot distant targets, {name_player} is clearly the superior bowman.",
-		str_scalp = " The scalp has pixels covering it.",
-		fn_effect = get_normal_attack(weapon_type = 'small_game'),
-		str_description = "It's a newly crafted minecraft bow, complete with a set of minecraft arrows",
-		acquisition = acquisition_smelting,
-		stat = stat_bow_kills,
-	),
-	EwWeapon(  # 23
-		id_weapon = weapon_id_dclaw,
-		alias = [
-			"dragon claw",
-		],
-		str_crit = "{name_player} runs like a madman towards {name_target}, {name_target} swings but is deftly parried by {name_player}, {name_player} hoists their dragon claw into the air and ripostes {name_target} for massive damage ***!!!Critical Hit!!!***",
-		str_miss = "{name_player} swings but {name_target} is in the middle of a dodge roll and is protected by iframes. **!!Miss!!**",
-		str_equip = "You place the core of the dragon claw on your hand and it unfolds around it, conforming to the contour of your hands, claws protude out the end of your fingers as your hand completes its transformation into the *dragon claw*.",
-		str_name = "dragon claw",
-		str_weapon = "a dragon claw",
-		str_weaponmaster_self = "You are a rank {rank} master of the dragon claw.",
-		str_weaponmaster = "They are a rank {rank} master of the dragon claw.",
-		#str_trauma_self = "Three smoldering claw marks are burned into your flesh, the flames `won't seem to extinguish.",
-		#str_trauma = "Three smoldering claw marks are burned into their flesh, the flames won't seem to extinguish.",
-		str_kill = "***Thwip.*** {name_player}'s dragon claw cuts the air followed by a trail of flame and blood, the camera pans out and {name_target} is shown, cut in twain. {emote_skull}",
-		str_killdescriptor = "cut to pieces",
-		str_damage = random.choice(["{name_target} is slashed across the {hitzone}!!","{name_player} furiously slashes {name_target} across the {hitzone}!!","{name_player} flicks their fingers and a jet of flame ignites from the dragon claw, burning {name_target} in the {hitzone}!!"]),
-		str_duel = "**SLICE!! SWIPE!! SLASH!!** {name_player} and {name_target} cut the fuck out of eachother, a fire extinguisher is never more than a meter away.",
-		str_scalp = "The scalp is burning and doesn't look like it's gonna stop.",
-		fn_effect = get_normal_attack(weapon_type = 'incendiary'),
-		str_description = "It's the core of a Dragon Claw, it will morph around whatever hand it is held by granting them the power of the elusive GREEN EYES SLIME DRAGON. If you listen closely you can hear whines of the dragon soul as it remains perpetually trapped in the weapon.",
-		acquisition = acquisition_smelting,
-		stat = stat_dclaw_kills,
-		classes = [weapon_class_burning, weapon_class_captcha],
-		captcha_length = 4
-	),
-
-	EwWeapon(  # 24
-		id_weapon=weapon_id_spraycan,
-		alias=[
-			"spray can",
-			"spray"
-		],
-		str_crit="**Critical hit!!** {name_player} flicks the nozzle off their spray can and lights it like a fuse! {name_target} gets nasty burns and a fresh coat of paint! **WHOOSH!!!**",
-		str_miss="**MISS!!** {name_player} attempts a spray attack, but the wind blows it back in their face!",
-		str_equip="You hold the spray can tightly, hoping to god somebody confuses it for a gun.",
-		str_name="spray can",
-		str_weapon="a spray can",
-		str_weaponmaster_self="You are a rank {rank} vandal of the spray can.",
-		str_weaponmaster="They are a rank {rank} vandal of the spray can.",
-		# str_trauma_self = "You're having trouble breathing, and the inside of your mouth is off-color.",
-		# str_trauma = "They're weirdly short of breath, and their mouth and tongue are off-color.",
-		str_kill="***PPPPPPSSSSSSSSSHHHHHhhhhhfff.*** {name_player} forcibly opens {name_target}'s mouth and sprays everything they have into their lungs. Their eyes roll back into their head and, trembling, they slowly asphyxiate in your arms. {emote_skull}",
-		str_killdescriptor="suffocated",
-		str_damage=random.choice(["{name_target} is whacked across the {hitzone}!!",
-								  "{name_player} sprays {name_target} with paint, making them a gaudy color in the {hitzone}!!",
-								  "{name_player} humiliates {name_target} by bringing a spray can to a gunfight, mentally damaging them in the {hitzone}!!"]),
-		str_duel="**PSSS PSSS PSSSSSHH!** {name_player} and {name_target} spray the dojo walls until they get dizzy from the smell.",
-		str_scalp="The scalp is a nice shade of mauve.",
-		fn_effect = get_normal_attack(weapon_type = 'tool'),
-		str_description="It's a Based Hardware brand spray can, in your gang's color. The blurb on the backside preaches about the merits of street art and murals, but you're pretty sure that's just to cover their ass.",
-		vendors = [vendor_basedhardware, vendor_breakroom],
-		stat=stat_spraycan_kills,
-		classes=[weapon_class_paint, weapon_class_captcha],
-		#sap_cost=2,
-		captcha_length=3,
-		is_tool = 1,
-		# str_backfire = "As {name_player} shakes the can to fire another shot, the thing suddenly explodes on them!",
-		tool_props = {
-		'reg_spray' : "You run down the streets, tagging buildings, street signs and old ladies with spray paint in the image of the {gang}!",
-		'miss_spray' : "**Miss!** Your can seems to be low on spray. You fill it up and give it a good shake. Good as new!",
-		'crit_spray' : "**Critical hit!** You dual wield spray cans, painting an urban masterpiece in one hand and shooting toxic chemicals into a cop's mouth with the other!",
-		'equip_spray' : "You get your trusty spray paint at the ready."}
-		),
-	EwWeapon(  # 25
-		id_weapon=weapon_id_paintgun,
-		alias=[
-			"paint gun",
-			"splatoon"
-		],
-		str_crit="**Critical hit!!** {name_player} aims down the sights with the precision of a video game real life sniper, shooting {name_target} in the eyes from 30 yards! **SPLAAAAAT!!!**",
-		str_miss="**MISS!!** {name_player} fires off a volley of paint, but {name_target} jumps behind cover!",
-		str_equip="Now listen here. You just equipped a paint gun. Keep in mind this is the weapon that boomer families shoot each other with to have fun. Enjoy trying to kill with it.",
-		str_name="paint gun",
-		str_weapon="a paint gun",
-		str_weaponmaster_self="You are a rank {rank} vandal of the paint gun.",
-		str_weaponmaster="They are a rank {rank} vandal of the paint gun.",
-		# str_trauma_self = "You have a splitting headache.",
-		# str_trauma = "They look hungover, almost like their entire body exploded.",
-		str_kill="***SPLAAAAART!!!!*** {name_player} fatally strikes {name_target}, and they explode from the inside out! There's a lot more gore than when you see it happen in Splatoon, though.{emote_skull}",
-		str_killdescriptor="imploded",
-		str_damage=random.choice(["{name_target} is splatted in the {hitzone}!!",
-								  "{name_player} shoots {name_target} with paint, making them a gaudy color in the {hitzone}!!",
-								  "{name_player} attacks {name_target} with harmless paint!!"]),
-		str_duel="**SPLAT TAT TAT!!** {name_player} and {name_target} harass everyone in the dojo with their paint guns.",
-		str_scalp="The scalp is colorful, from both blood and paint.",
-		fn_effect = get_normal_attack(weapon_type = 'tool'),
-		str_description="It's an industrial strength two handed paint gun with a sniper scope attached. What do they use this for in industry, anyway?",
-		vendors=[vendor_basedhardware, vendor_breakroom],
-		stat=stat_paintgun_kills,
-		classes=[weapon_class_paint, weapon_class_ammo, weapon_class_captcha],
-		clip_size = 15,
-		price = 5000,
-		#sap_cost=2,
-		captcha_length=4,
-		is_tool = 1,
-		str_reload = "*Click.* You grab a paint cylinder from god knows where and load it into your gun, chucking the leftover one behind an alleyway.",
-		# str_backfire = "Whoops, looks like somebody didn't fasten the paint cylinder hard enough! {name_player} gets a thorough spray to the face!",
-		tool_props = {
-			'reg_spray':  "You find a patch of wall several yards away that hasn't been vandalized yet. Time to take aim and...BAM! Nice shot!",
-			'miss_spray': "**Miss!** Your aim was as sharp as ever, but a fucking pigeon took the hit! Christ, what are the odds?",
-			'crit_spray': "**Critical hit!** The paint bullet skids a wall, spreading your paint across the whole thing!",
-			'equip_spray': "You load a clip of paint into the gun and throw it onto your back, kinda like Rambo if he were an art major."
-		}
-	),
-	EwWeapon(  # 26
-		id_weapon=weapon_id_paintroller,
-		alias=[
-			"paint roller",
-			"roller"
-		],
-		str_crit="**Critical hit!!** {name_player}  knocks {name_target} to the ground and does a golf swing to their vulnerable little head, sending them spinning. **FWAP!!!**",
-		str_miss="**MISS!!** {name_player} does cringey bo staff jujitsu moves with the roller and forgets to actually attack {name_target}!",
-		str_equip="You hold the paint roller in your hand. The light plastic broom handle and spongy brush are sure to deal at least 10 damage.",
-		str_name="paint roller",
-		str_weapon="a paint roller",
-		str_weaponmaster_self="You are a rank {rank} vandal of the paint roller.",
-		str_weaponmaster="They are a rank {rank} vandal of the paint roller.",
-		# str_trauma_self = "There's a gaudy colored dent in your skull.",
-		# str_trauma = "There is a gaudy colored dent in their skull.",
-		str_kill="***CA-CRACK!*** {name_player} opens {name_target}'s skull like an egg using the dull metal edge of the roller. It appears to be hollow, after all, {name_target} was stupid enough to get killed with a fucking paint roller.{emote_skull}",
-		str_killdescriptor="cracked open",
-		str_damage=random.choice(["{name_target} is swatted in the {hitzone}!!",
-								  "{name_player} slaps {name_target} with paint, making them a gaudy color in the {hitzone}!!",
-								  "{name_player} rolls paint all over {name_target}'s {hitzone}!!"]),
-		str_duel="{name_player} and {name_target} quietly pass the time rolling paint over the windows of nearby houses You both have learned tranquility.",
-		str_scalp="The scalp is split in half, with a big hole right in the middle.",
-		fn_effect = get_normal_attack(weapon_type = 'tool'),
-		price = 4500,
-		str_description="It's a long, broom-like plastic paint roller with a spongy brush and metal axle. The modern man's bo staff.",
-		vendors=[vendor_basedhardware, vendor_breakroom],
-		stat=stat_paintroller_kills,
-		classes=[weapon_class_paint, weapon_class_captcha],
-		#sap_cost=2,
-		captcha_length=4,
-		is_tool=1,
-		# str_backfire = "{name_player} waves the paint roller around like it's a plastic toy, spreading paint nowhere but giving themselves a thorough welt in the head from the 2 square inches of it that could actually do any damage. How'd they manage that?",
-		tool_props = {
-			'reg_spray' : "You roll paint over as much surface area as your puny little Juvie legs can take you to.",
-			'miss_spray' : "**Miss!** The sponge on your roller snaps off and it takes too long for you to notice. What a waste!",
-			'crit_spray' : "**Critical hit!** Your mind goes blank in a painting-induced rage. When you wake up, all your surroundings are {color} You should do that more often!",
-			'equip_spray' : "You grab your paint roller and strap it on your back."}
-		),
-	EwWeapon(  # 27
-		id_weapon=weapon_id_paintbrush,
-		alias=[
-			"paint brushes",
-			"brush"
-		],
-		str_crit="**Critical hit!!** {name_player}  stabs {name_target} with one brush and paints over their eyes with another!  **HOT DOG!!!**",
-		str_miss="**MISS!!** {name_player} throws the brushes at {name_target}, but they get hit with the soft bristles instead of the pointy bit!",
-		str_equip="If only you had a whittling knife that could sharpen paintbrush handles. That way you could equip the knife as a weapon instead of this.",
-		str_name="paintbrushes",
-		str_weapon="paintbrushes",
-		str_weaponmaster_self="You are a rank {rank} vandal of the paintbrush.",
-		str_weaponmaster="They are a rank {rank} vandal of the paintbrush.",
-		# str_trauma_self = "You have bruises all over your body and you can't get the paint out of your clothes.",
-		# str_trauma = "They have bruises all over their body, and they can't get the paint out of their clothes.",
-		str_kill="***MASTERPIECE!*** {name_target} takes a mortal brush to the forehead, courtesy of {name_player}'s talent as a painter. {emote_skull}",
-		str_killdescriptor="paintbrushed to death",
-		str_damage=random.choice(["{name_target} is handlestabbed in the {hitzone}!!",
-								  "{name_player} flecks {name_target} with paint, making them a gaudy color in the {hitzone}!!",
-								  "{name_player} grazes {name_target}'s {hitzone} with coarse bristles!!"]),
-		str_duel="{name_player} and {name_target} paint random text commands on the walls outside the Dojo. {name_target} paints some furry art when nobody's looking.",
-		str_scalp="The scalp has a bunch of welts, and has a faint smell of lead.",
-		fn_effect = get_normal_attack(weapon_type = 'tool'),
-		str_description="A stack of large, coarse-bristled paintbrushes, linked together on a burlap string.",
-		vendors=[vendor_basedhardware, vendor_breakroom],
-		stat=stat_paintbrush_kills,
-		classes=[weapon_class_paint, weapon_class_captcha],
-		#sap_cost=2,
-		price = 100,
-		captcha_length=3,
-		is_tool=1,
-		# str_backfire = "In an attempt to paint faster, {name_player} sticks one of the handles in their mouth and try to use it to cover more ground. Instead, they broke your teeth and scraped their cheek on a hard brick surface. Better not try that again...",
-		tool_props = {
-			'reg_spray' : "You paint vulgar {gang} symbols on as many buildings as you can.",
-			'miss_spray' : "**Miss!** You finish with a paint can and have to switch! You waste too much time getting the can open.",
-			'crit_spray' : "**Critical hit!**  You hold the paint can in your mouth and start crab walking, throwing paint along the wall as you do it! Somehow, this is more efficient!",
-			'equip_spray' : "You get your brushes at the ready."}
-	),
-	EwWeapon(  # 28
-		id_weapon=weapon_id_watercolors,
-		alias=[
-			"paint brushes",
-			"brush"
-		],
-		str_crit="```css\n\"oooOOOOOOOH LA LA! {name_target} is exposed to {name_player}'s watercolor pornography! They won't be able to recover from that!\"\n```",
-		str_miss="```css\n[{name_player} paints a picture for {name_target}. It does no damage, as expected.]\n```",
-		str_equip="```ini\n[You get a nice mug to dip your little paintbrush in, and open your 12 set of watercolors. Look out world, here comes you!]\n```",
-		str_name="watercolors",
-		str_weapon="a set of watercolors",
-		str_weaponmaster_self="You are a rank {rank} flaming homosexual of watercolors.",
-		str_weaponmaster="They are a rank {rank} flaming homosexual of watercolors.",
-		# str_trauma_self = "You are eternally humiliated after being murdered by a gangster wielding watercolor paints.",
-		# str_trauma = "They are eternally humiliated after being murdered by a gangster wielding watercolor paints.",
-		str_kill="```bash\n\"HUUUUUUH?? {name_target} goes and kills themselves after having an existential crisis! {name_player} seems to have done this with only their own retardation!\"\n```",
-		str_killdescriptor="driven to suicide",
-		str_damage="```ini\n[{name_player} paints a picture of {name_target}. Their self esteem takes a hit!]\n```",
-		str_duel="```json\n\"{name_player} and {name_target} practice art using Dojo-owned easels and canvases. Eventually, the training session breaks down and, you just throw paint water at each other and giggle like schoolgirls.\"\n```",
-		str_scalp="The scalp is perfectly intact.",
-		fn_effect = get_normal_attack(weapon_type = 'tool'),
-		str_description="A 12 pack of watercolors, the kind you used when you were a 5 years old boy.",
-		vendors=[vendor_basedhardware, vendor_breakroom],
-		stat=stat_watercolor_kills,
-		classes=[weapon_class_paint, weapon_class_captcha],
-		#sap_cost=2,
-		price = 1300,
-		captcha_length=3,
-		is_tool=1,
-		# str_backfire = "{name_player} has the idea of trying to paint their parents fucking, thinking it will be really funny and everyone will love them. Pretty soon we're going to have to ban watercolors because people like you are using them to molest yourself.",
-		tool_props = {
-			'reg_spray' : "Nice drawing, {curse}! ",
-			'miss_spray' : "**Miss!** Your painting sucks. God, you're stupid. ",
-			'crit_spray' : "After the thousandth failed watercolor gesamtkunstwerk you decide enough is enough. Fuck this. Fuck the gangs, fuck the violence, fuck the perpetually rotting lets player that compels you to rigor mortis yourself more frequently than you eat breakfast. The spite is so concentrated that it compels you to turn your life around. You get a fake ID, join the PTA, and rope them into cleaning every last inch of this district until the homeless population smell like citrus and give out free, non-tainted lollipops. However, your newfound peaceful life is interrupted by the night terrors ENDLESS WAR now gives you on a daily basis, and you decide to go back to being a gangster. You suppose some things never change.",
-			'equip_spray' : "You get out your 12 pack of watercolors. Can't believe you have to use one of these."
-		}
-	),
-	EwWeapon(  # 29
-		id_weapon=weapon_id_thinnerbomb,
-		alias=[
-			"thinner",
-			"thinnerbombs"
-		],
-		str_crit="**Critical hit!!** {name_player} slams {name_target} with a bottle of paint thinner, showering their face with broken glass and getting some of the thinner down their gullet. They fall back, dazed and bleeding.",
-		str_miss="**MISS!!** {name_player} is too dazed by their own chemicals to make a move! They drop the bottle on accident, throwing vapors all over the place.",
-		str_equip="You pull out the thinner bombs and hold their bottlenecks between your fingers. Never has a not-weapon ever felt so cool.",
-		str_name="thinner bombs",
-		str_weapon="thinner bombs",
-		str_weaponmaster_self="You are a rank {rank} vandal of the thinner bomb.",
-		str_weaponmaster="They are a rank {rank} vandal of the thinner bomb.",
-		# str_trauma_self = "You have the hangover from hell.",
-		# str_trauma = "They have the hangover from hell.",
-		str_kill="***WHAT A SIZZLER!*** {name_target}, dazed from the concentrated toxic chemicals in the air, falls to the ground, giving {name_player} the chance to stab them through the neck with the broken bottle. Inhalants. Not even once. {emote_skull}",
-		str_killdescriptor="drugged",
-		str_damage=random.choice(["{name_target} gets a thinnerbomb to the {hitzone}!!",
-								  "{name_player} slashes {name_target} with a broken thinnerbomb! Ooh, right in the {hitzone}!!"]),
-		str_duel="{name_player} and {name_target} build a resistance to the noxious chemicals they're using by drinking paint thinner together. Cheers.",
-		str_scalp="The scalp smells awful, you can hardly hold it.",
-		fn_effect = get_normal_attack(weapon_type = 'tool'),
-		str_description="A pack of brittle glass bottles filled with paint thinner. This stuff vaporizes like nobody's business, and could strip the osmotic membrane off a slimeoid.",
-		vendors=[vendor_basedhardware, vendor_breakroom],
-		stat=stat_thinnerbomb_kills,
-		classes=[weapon_class_paint, weapon_class_captcha],
-		#sap_cost=2,
-		price = 15000,
-		captcha_length=4,
-		is_tool = 1,
-		# str_backfire = "You haven't had a good buzz in awhile, so you take a whiff of one of your thinner bombs. Great trip and all, but you rough yourself up convulsing on the ground while it happens.",
-		tool_props = {
-		'reg_spray' : "You find a vandalized wall and toss a thinner bomb on it! You hear a faint sizzling as paint begins to strip off the walls. Sick!",
-		'miss_spray' : "**Miss!** You make a mistake on the throw's distance and it bursts uselessly on the ground. You got to do some littering, so at least there's that.",
-		'crit_spray' : "**Critical hit!** You take out a paint bomb and throw it at a particularly fragile looking building. The chemicals you used were so caustic that they burned a hole through the whole wall, preventing anyone from painting it for all of time!",
-		'equip_spray' : "You get your glass thinner bombs out you you can throw them in a moment's notice."
-	}),
-	EwWeapon( # 30
-		id_weapon = weapon_id_staff,
-		alias = [
-			"eldritchstaff",
-			"spookystaff",
-			"reprehensiblerod",
-			"wickedwand",
-			"frighteningfaggot"
-		],
-		str_miss = "Your mind goes blank as you feel slime disappear from your body in preparation for a deadly attack.",
-		str_damage = "{name_player} finalizes their invocation. " + random.choice([
-			"Gravity violently increases in the space around {name_target}, slamming them into the ground.", 
-			"A blinding white light shines from {name_target}'s {hitzone} as it burns hotter than the surface of the sun.", 
-			"Spectral hands caress {name_target}'s body, leaving gaping wounds in their path.", 
-			"An unseen force suddenly yoinks {name_target} by their {hitzone}, sending them flying into the air.",
-			"A pitch black horror forms around {name_target}'s {hitzone} and tears into it."
-		]),
-		str_crit = "{name_player} notices {name_target} still recoiling from the damage, and takes the chance to bonk the everliving shit out of them with their staff. **Critical hit!!**",
-		str_kill = "A mass of tiny hands erupts from the ground below {name_target}, grabbing on to their body. Their screams echo across the streets as they're dragged through the ground and into the sewers.",
-		str_equip = "You equip the eldritch staff.",
-		str_name = "eldritch staff",
-		str_weapon = "an eldritch staff",
-		str_weaponmaster_self = "You are a rank {rank} conduit of the ones below.",
-		str_weaponmaster = "They are a rank {rank} conduit of the ones below.",
-		str_killdescriptor = "cast down",
-		str_duel = "{name_player} and {name_target} compare notes on their understanding of the eldritch fuckery they've each experienced.",
-		str_description = "An intricate wooden staff with a cloudy crystal on its handle. It looks fucking class, but it also gives you the creeps.",
-		str_scalp = "It's covered in symbols written with a strange black substance.",
-		fn_effect = wef_staff,
-		acquisition = acquisition_smelting,
-		stat = stat_staff_kills,
-		#sap_cost = 2,
-		captcha_length = 10,
-	),
-	EwWeapon( # 31
-		id_weapon = weapon_id_hoe,
-		str_miss = "**MISS!!** {name_player}'s hoe strikes the earth with a loud THUD.",
-		str_damage = "{name_player} scrapes their hoe across {name_target}'s {hitzone}.",
-		str_crit = "**CRITICAL HIT!!** {name_player} gets their hoe deep into {name_target}'s body, cutting up their vitals!",
-		str_kill = "{name_player} pushes {name_target} to the ground. After an intense windup, they slam their hoe down on {name_target}'s neck, decapitating them in the process.",
-		str_equip = "You ready your hoe.",
-		str_name = "hoe",
-		str_weapon = "a hoe",
-		str_weaponmaster_self = "You are a rank {rank} farmer.",
-		str_weaponmaster = "They are a rank {rank} farmer.",
-		str_killdescriptor = "!reaped",
-		str_duel = "{name_player} and {name_target} discuss their latest harvest and exchange farming tips.",
-		str_description = "It's a farming hoe.",
-		str_scalp = "It's covered in dirt.",
-		fn_effect = get_normal_attack(weapon_type = 'tool'),
-		price = 100000,
-		vendors = [vendor_atomicforest],
-		classes = [weapon_class_farming, weapon_class_juvie],
-		stat = stat_hoe_kills,
-		#sap_cost = 2,
-		captcha_length = 2,
-		is_tool = 1,
-	),
-	EwWeapon( # 32
-		id_weapon = weapon_id_pitchfork,
-		str_miss = "**MISS!!** {name_player}'s pitchfork is planted firmly into the ground.",
-		str_damage = "{name_player} stabs {name_target}'s {hitzone} with their pitchfork!",
-		str_crit = "**CRITICAL HIT!!** {name_player} pokes several holes in {name_target}!",
-		str_kill = "{name_player} plants their pitchfork firmly into {name_target} and lifts them high into the air. After {name_target} loses consciousness, {name_target} throws them to the ground.",
-		str_equip = "You pick up your pitchfork and give the ground a light tap with the handle's end.",
-		str_name = "pitchfork",
-		str_weapon = "a pitchfork",
-		str_weaponmaster_self = "You are a rank {rank} farmer.",
-		str_weaponmaster = "They are a rank {rank} farmer.",
-		str_killdescriptor = "!reaped",
-		str_duel = "{name_player} and {name_target} joust with their pithforks. Thankfully, no one gets hurt in the process.",
-		str_description = "It's a farming pitchfork.",
-		str_scalp = "It's got three holes in it.",
-		fn_effect = get_normal_attack(weapon_type = 'tool'),
-		price = 100000,
-		vendors = [vendor_atomicforest],
-		classes = [weapon_class_farming, weapon_class_juvie],
-		stat = stat_pitchfork_kills,
-		#sap_cost = 2,
-		captcha_length = 2,
-		is_tool = 1,
-	),
-	EwWeapon( # 33 
-		id_weapon = weapon_id_shovel,
-		str_miss = "**MISS!!** {name_player}'s shovel is planted firmly into the ground.",
-		str_damage = "{name_player} swings their shovel at {name_target}'s {hitzone}!",
-		str_crit = "**CRITICAL HIT!** The flat end of {name_player}'s shovel impacts {name_target}'s chest! They start coughing up blood!",
-		str_kill = "*BONK!* {name_player}'s shovel lands right on top of {name_target}'s head. Their skull and brain is completely crushed by the impact. {name_player} buries them in a shallow grave.",
-		str_equip = "You grip your shovel tightly in both hands.",
-		str_name = "shovel",
-		str_weapon = "a shovel",
-		str_weaponmaster_self = "You are a rank {rank} farmer.",
-		str_weaponmaster = "They are a rank {rank} farmer.",
-		str_killdescriptor = "!digged",
-		str_duel = "{name_player} and {name_target} perform a high-shovel. The moment could not be more perfect.",
-		str_description = "It's a shovel.",
-		str_scalp = "It's flattened.",
-		fn_effect = get_normal_attack(weapon_type = 'tool'),
-		price = 100000,
-		vendors = [vendor_atomicforest],
-		classes = [weapon_class_juvie],
-		stat = stat_shovel_kills,
-		#sap_cost = 2,
-		captcha_length = 2,
-		is_tool = 1,
-	),
-	EwWeapon( # 34 
-		id_weapon = weapon_id_slimeringcan,
-		str_miss = "**MISS!!** Spouts of slime from {name_players} Slimering Can fly everywhere!",
-		str_damage = "{name_player} pours slime onto {name_target}'s {hitzone}. What the fuck is that going to accomplish?",
-		str_crit = "**CRITIAL HIT!!** {name_player} pours slime onto {name_target}'s eyes! How unsanitary!",
-		str_kill = "{name_player} rams their Slimering Can down {name_target}'s throat. {name_target} chokes to death on slime.",
-		str_equip = "You pick up your Slimering Can.",
-		str_name = "slimering can",
-		str_weapon = "a slimering can",
-		str_weaponmaster_self = "You are a rank {rank} green thumbed coward.",
-		str_weaponmaster = "They are a rank {rank} green thumbed coward.",
-		str_killdescriptor = "drowned",
-		str_duel = "{name_player} and {name_target} water flowers together. Sometimes it's nice to be a fucking weak willed coward, y'know?",
-		str_description = "It's a slimering can.",
-		str_scalp = "It's soaking wet.",
-		fn_effect = get_normal_attack(weapon_type = 'tool'),
-		price = 100000,
-		vendors = [vendor_atomicforest],
-		classes = [weapon_class_juvie],
-		stat = stat_slimeringcan_kills,
-		#sap_cost = 2,
-		captcha_length = 2,
-		is_tool = 1,
-	),
-    EwWeapon(  # 35
-        id_weapon=weapon_id_fingernails,
-        alias=[
-            "fingernails",
-            "nails"
-        ],
-        str_crit="**Critical hit!!** {name_target} is cut deep!!",
-        str_miss="",
-        str_equip="",
-        str_name="fingernails",
-        str_weapon="their fingernails",
-        str_weaponmaster_self="",
-        str_weaponmaster="",
-        # str_trauma_self = "A single clean scar runs across the entire length of your body.",
-        # str_trauma = "A single clean scar runs across the entire length of their body.",
-        str_kill="{name_player} lunges at {name_target} with fingernails bared! They're mercilessly rips them to pieces, tufts of skin flying every which way! **BRRRRRRRAP!!!** When the dust settles, {name_target} is unrecognizable, and more importantly, dead as fuck. {emote_skull}",
-        str_killdescriptor="torn apart",
-        str_damage="{name_target} is slashed across the {hitzone}!!",
-        str_duel="",
-        str_description="",
-        str_scalp=" Multiple slash marks run across it.",
-        fn_effect=wef_fingernails,
-        price=0,
-        vendors=[],
-        classes=[],
-        stat=stat_fingernails_kills,
-        # sap_cost = 3,
-        captcha_length=8
-    ),
-	EwWeapon(  # 35
-		id_weapon=weapon_id_roomba,
-		alias=[
-			"roomba",
-			"vaccuum"
-		],
-		str_crit="**Critical hit!!** {name_target} gets a concussion via roomba to the face!!",
-		str_miss="**MISS!** The roomba forgets where it is and begins tripping around!",
-		str_equip="You turn on your Roomba and place it on the ground.",
-		str_name="roomba",
-		str_weapon="a roomba",
-		str_weaponmaster_self="You are a rank {rank} roomba acolyte.",
-		str_weaponmaster="They are a rank {rank} roomba acolyte.",
-		# str_trauma_self = "A single clean scar runs across the entire length of your body.",
-		# str_trauma = "A single clean scar runs across the entire length of their body.",
-		str_kill="{name_player} jumps at {name_target} using the roomba as a springboard! Screaming bloody murder, they beat {name_target} to within an inch of their life. The roomba follows, sucking whatever morsel of slime was left. Job finished. {emote_skull}",
-		str_killdescriptor="sucked dry",
-		str_damage="{name_player}'s roomba sucks gobs of slime out of {name_target}'s {hitzone}!!",
-		str_duel="{name_player} and {name_target} begin engineering their portable vaccums into high class battle bots. By the time you're done the Dojo floor is spotless and everyone nearby is dead.",
-		str_description="It's a high powered portable vaccuum designed to clean up dust. You use it to spread paint around by attaching a spray can to the back.",
-		str_scalp=" It looks stretched and wrinkled.",
-		fn_effect = get_normal_attack(weapon_type = 'tool'),
-		price=40000,
-		# str_backfire="You roomba turns on you! Its shitty AI thinks your feet are its prey, and it sucks away some precious slime!",
-		vendors=[vendor_basedhardware],
-		classes=[weapon_class_paint, weapon_class_captcha],
-		stat=stat_fingernails_kills,
-		# sap_cost = 3,
-		captcha_length=8,
-		tool_props = {
-		'reg_spray' : "The roomba continues its intrepid journey spraying paint around town.",
-		'miss_spray' : "**Miss!** Fuck, the thing got stuck on a pothole again.",
-		'crit_spray' : "**Critical hit!** A bystander walking by kicks your roomba as it's moving, which inadvertently overclocks its processor!! It speeds around the area with reckless abandon. Go go go!",
-		'equip_spray' : "You pull out your roomba and set it on the ground."
-	}
-	),
-	EwWeapon(  # 36
-		id_weapon=weapon_id_laywaster,
-		alias=[
-			"chainsaw",
-			"megachainsaw",
-			"widowmaker",
-			"jessica"
-		],
-		str_crit="**Critical Hit!** {name_player} snaps {name_target} between two of the sawblades, ripping mercilessly into flesh and nearly vaporizing the spraying blood!",
-		str_miss="**Miss!** {name_player} swings the heavy blade around and hits nothing but air.",
-		str_equip="You rev up the Laywaster 9000.",
-		str_name="multiblade chainsaw",
-		str_weapon="a multiblade chainsaw",
-		str_weaponmaster_self="You are a rank {rank} master of the Laywaster 9000.",
-		str_weaponmaster="They are a rank {rank} master of the Laywaster 9000.",
-		# str_trauma_self = "You have a few large, gaping holes in your abdomen. Someone could stick their arm through the biggest one.",
-		# str_trauma = "They have a few large, gaping holes in your abdomen. Someone could stick their arm through the biggest one.",
-		str_kill="**VRRRRRRRRRRRRRRRRRRRRRRRRRRRR!**{name_player} swings violently through {name_target}'s unconscious body, each slash making them more unrecognizable than the last. As more and more blood flecks across {name_player}'s face, their opponent turns into a pile of viscera. {emote_skull}",
-		str_killdescriptor="shredded to a paste",
-		str_damage="{name_target}'s {hitzone} is torn into!! Blood flies everywhere!",
-		str_duel="**...** {name_player} and {name_target} clash with each other chainsaw blow for chainsaw blow like badasses.",
-		str_description="It's a multi-bladed chainsaw.",
-		str_scalp=" It's not really a scalp anymore, more of a paste.",
-		fn_effect=get_normal_attack(weapon_type='heavy'),
-		price=1000000,
-		vendors=[],
-		classes=[],
-		stat=stat_megachainsaw_kills,
-	),
-EwWeapon( # 37
-		id_weapon = weapon_id_chainsaw,
-		alias = [
-			"ripper",
-			"motoraxe"
-		],
-		str_crit = "**Critical hit!!** The jagged teeth of the chainsaw rest within {name_target}, body as the slime flies!!",
-		str_miss = "**You missed!!** In {name_player}’s excitement and desperation neither chain nor saw hits {name_target}!",
-		str_equip = "You equip the chainsaw.",
-		str_name = "chainsaw",
-		str_weapon = "a chainsaw",
-		str_weaponmaster_self = "You are a rank {rank} wielder of the chainsaw.",
-		str_weaponmaster = "They are a rank {rank} wielder of the chainsaw.",
-		#str_trauma_self = "Your body runs jagged with large chunks missing and patches of skin torn up.",
-		#str_trauma = "Their body runs jagged with large chunks missing and patches of skin torn up.",
-		str_kill = "**REEERNREERN!!** {name_player} revs up their chainsaw and carves up {name_target}’s torso, cutting through the guts,bile,viscera, and slime; sending it all flying. They’ve been cut down to size. {emote_skull}",
-		str_killdescriptor = "chainsaw’d",
-		str_damage = "The numerous finely tooth blades tear at {name_target}’s {hitzone}!!",
-		str_duel = "**...** {name_player} and {name_target} clash with each other chainsaw blow for chainsaw blow like badasses.",
-		str_description = "It's a chainsaw.",
-		#str_reload = "You desperately pull at the ripcord of your chainsaw trying to rev it back up to speed.",
-		#str_reload_warning = "**REEERNREERN--** *shhhhh…* **FUCK!!** {name_player}’s chainsaw just ran out of it’s rev!!",
-		str_scalp = "It’s more like a collection of dandruff then a scalp.",
-		fn_effect = get_normal_attack(weapon_type = 'heavy'),
-		price = 1000000,
-		vendors = [vendor_basedhardware],
-		classes= [],
-		stat = stat_chainsaw_kills,
-		captcha_length = 4
-	),
-
-]
-
-# A map of id_weapon to EwWeapon objects.
-weapon_map = {}
-
-# A list of weapon names
-weapon_names = []
 
 # Attacking type effects
 def atf_fangs(ctn = None):
@@ -6432,1226 +5016,6 @@ slimeoid_special_throw = "throw"
 slimeoid_special_TK = "TK"
 slimeoid_special_fire = "fire"
 slimeoid_special_webs = "webs"
-
-# All body attributes in the game.
-body_list = [
-	EwBody( # body 1
-		id_body = "teardrop",
-		alias = [
-			"tear",
-			"drop",
-			"oblong",
-			"a"
-		],
-		str_create = "You press a button on the body console labelled 'A'. Through the observation port, you see the rapidly congealing proto-Slimeoid take on a roughly teardrop-shaped form.",
-		str_body = "It is teardrop-shaped.",
-		str_observe = "{slimeoid_name} is bobbing its top-heavy body back and forth."
-	),
-	EwBody( # body 2
-		id_body = "wormlike",
-		alias = [
-			"long",
-			"serpent",
-			"serpentine",
-			"b"
-		],
-		str_create = "You press a button on the body console labelled 'B'. Through the observation port, you see the rapidly congealing proto-Slimeoid begin to stretch into an elongated form.",
-		str_body = "It is long and wormlike.",
-		str_observe = "{slimeoid_name} is twisting itself around, practicing tying its knots."
-	),
-	EwBody( # body 3
-		id_body = "spherical",
-		alias = [
-			"sphere",
-			"orb",
-			"ball",
-			"c"
-		],
-		str_create = "You press a button on the body console labelled 'C'. Through the observation port, you see the rapidly congealing proto-Slimeoid take on a roughly spherical form.",
-		str_body = "It is generally orb-shaped.",
-		str_observe = "{slimeoid_name} flops over onto one side of its round body."
-	),
-	EwBody( # body 4
-		id_body = "humanoid",
-		alias = [
-			"biped",
-			"human",
-			"d"
-		],
-		str_create = "You press a button on the body console labelled 'D'. Through the observation port, you see the rapidly congealing proto-Slimeoid curl into a foetal, vaguely humanoid form.",
-		str_body = "It is vaguely humanoid.",
-		str_observe = "{slimeoid_name} is scraping at something on the ground with its arms."
-	),
-	EwBody( # body 5
-		id_body = "tentacled",
-		alias = [
-			"squid",
-			"squidlike",
-			"tentacle",
-			"tentacles",
-			"e"
-		],
-		str_create = "You press a button on the body console labelled 'E'. Through the observation port, you see the rapidly congealing proto-Slimeoid begin to sprout long tendrils from its nucleus.",
-		str_body = "It is a mass of tendrils.",
-		str_observe = "{slimeoid_name} is moving its tentacles around, running them over one another."
-	),
-	EwBody( # body 6
-		id_body = "amorphous",
-		alias = [
-			"none",
-			"formless",
-			"f"
-		],
-		str_create = "You press a button on the body console labelled 'F'. Through the observation port, you see the rapidly congealing proto-Slimeoid accreting itself together with no distinct shape to speak of.",
-		str_body = "It has no defined shape.",
-		str_observe = "{slimeoid_name}'s body is spread out on the floor like a kind of living puddle."
-	),
-	EwBody( # body 7
-		id_body = "quadruped",
-		alias = [
-			"animal",
-			"g"
-		],
-		str_create = "You press a button on the body console labelled 'G'. Through the observation port, you see the rapidly congealing proto-Slimeoid beginning to grow bones and vertebrae as it starts to resemble some kind of quadruped.",
-		str_body = "It has a body shape vaguely reminiscent of a quadruped.",
-		str_observe = "{slimeoid_name} has its hindquarters lowered in a sort of sitting position."
-	)
-]
-
-# A map of id_body to EwBody objects.
-body_map = {}
-
-# A list of body names
-body_names = []
-
-# Populate body map, including all aliases.
-for body in body_list:
-	body_map[body.id_body] = body
-	body_names.append(body.id_body)
-
-	for alias in body.alias:
-		body_map[alias] = body
-
-# All head attributes in the game.
-head_list = [
-	EwHead( # head 1
-		id_head = "eye",
-		alias = [
-			"cyclops",
-			"a"
-		],
-		str_create = "You press a button on the head console labelled 'A'. Through the observation port, you see a dark cluster within the proto-Slimeoid begin to form into what looks like a large eye.",
-		str_head = "Its face is a single huge eye.",
-		str_feed = "{slimeoid_name} swallows the {food_name} whole.",
-		str_fetch = "You toss a ball for {slimeoid_name} to fetch. {slimeoid_name}'s huge eye follows the ball's arc, and it makes a leap to catch it!!\n\n...only to get socked right in the eye, sending it to the floor in pain. Depth perception... it's truly a gift."
-	),
-	EwHead( # head 2
-		id_head = "maw",
-		alias = [
-			"mouth",
-			"b"
-		],
-		str_create = "You press a button on the head console labelled 'B'. Through the observation port, you see an opening form in what you think is the proto-Slimeoid's face, which begins to sprout large pointed teeth.",
-		str_head = "Its face is a huge toothy mouth.",
-		str_feed = "{slimeoid_name} crunches the {food_name} to paste with its huge teeth.",
-		str_fetch = "You toss a ball for {slimeoid_name} to fetch. {slimeoid_name} slavers and drools as it awaits the throw, and when it sees the ball start to fly, it lunges to grab it out of the air in its huge toothy maw, slicing it to shreds with its teeth in seconds."
-	),
-	EwHead( # head 3
-		id_head = "void",
-		alias = [
-			"hole",
-			"c"
-		],
-		str_create = "You press a button on the head console labelled 'C'. Through the observation port, you see what you thought was the proto-Slimeoid's face suddenly sucked down into its body, as though by a black hole.",
-		str_head = "Its face is an empty black void.",
-		str_feed = "The {food_name} disappears into the unknowable depths of {slimeoid_name}'s face hole.",
-		str_fetch = "You toss a ball for {slimeoid_name} to fetch. {slimeoid_name} positions itself to catch the ball in it's... face? The ball falls into the empty void of {slimeoid_name}'s face, then just keeps falling, falling, falling, down into the depths, falling so far it disappears forever."
-	),
-	EwHead( # head 4
-		id_head = "beast",
-		alias = [
-			"animal",
-			"dragon",
-			"d"
-		],
-		str_create = "You press a button on the head console labelled 'D'. Through the observation port, you see the beginnings of an animal-like face forming on your proto-Slimeoid, with what might be eyes, a nose, teeth... maybe.",
-		str_head = "Its face is that of a vicious beast.",
-		str_feed = "{slimeoid_name} gobbles up the {food_name} greedily.",
-		str_fetch = "You toss a ball for {slimeoid_name} to fetch. {slimeoid_name} focuses its eyes and bares its teeth, then makes a flying leap, snapping the ball clean out of the air in its jaws! It comes back to you and drops the ball at your feet. Good boy!"
-	),
-	EwHead( # head 5
-		id_head = "insect",
-		alias = [
-			"bug",
-			"insectoid",
-			"e"
-		],
-		str_create = "You press a button on the head console labelled 'E'. Through the observation port, you see the proto-Slimeoid suddenly bulge with a series of hard orbs which congeal into what appear to be large compound eyes.",
-		str_head = "It has bulging insectoid eyes and mandibles.",
-		str_feed = "{slimeoid_name} cuts the {food_name} into pieces with its mandibles.",
-		str_fetch = "You toss a ball for {slimeoid_name} to fetch. {slimeoid_name} stands motionless, presumably watching the ball with its large compound eyes, before darting toward it as it sails through the air, snapping its mandibles around the ball and slicing it cleanly in two."
-	),
-	EwHead( # head 6
-		id_head = "skull",
-		alias = [
-			"skeleton",
-			"f"
-		],
-		str_create = "You press a button on the head console labelled 'F'. Through the observation port, you see the proto-Slimeoid's frontal features twist into a ghastly death's-head.",
-		str_head = "Its face resembles a skull.",
-		str_feed = "{slimeoid_name} spills half the {food_name} on the floor trying to chew it with its exposed teeth.",
-		str_fetch = "You toss a ball for {slimeoid_name} to fetch. {slimeoid_name} moves as quickly as it can to make the catch, making it just before the ball lands! With a thunk, the ball lodges itself in {slimeoid_name}'s open eye socket. {slimeoid_name} yanks it out and tosses the ball back to you. Euughh."
-	),
-	EwHead( # head 7
-		id_head = "none",
-		alias = [
-			"g"
-		],
-		str_create = "You press a button on the head console labelled 'G'. Through the observation port, you see the proto-Slimeoid's front end melt into an indistinct mass.",
-		str_head = "It has no discernable head.",
-		str_feed = "{slimeoid_name} just sort of... absorbs the {food_name} into its body.",
-		str_fetch = "You toss a ball for {slimeoid_name} to fetch. {slimeoid_name} moves under the ball as it flies through the air, but makes no attempt to catch it in its mouth on account of having none. The ball lands next to {slimeoid_name}, who merely looks on. Actually, you can't tell where it's looking."
-	)
-]
-
-# A map of id_head to EwBody objects.
-head_map = {}
-
-# A list of head names
-head_names = []
-
-# Populate head map, including all aliases.
-for head in head_list:
-	head_map[head.id_head] = head
-	head_names.append(head.id_head)
-
-	for alias in head.alias:
-		head_map[alias] = head
-
-# All mobility attributes in the game.
-mobility_list = [
-	EwMobility( # mobility 1
-		id_mobility = "legs",
-		alias = [
-			"animal",
-			"quadruped",
-			"biped",
-			"jointed",
-			"limbs",
-			"a"
-		],
-		str_advance = "{active} barrels toward {inactive}!",
-		str_retreat = "{active} leaps away from {inactive}!",
-		str_advance_weak = "{active} limps toward {inactive}!",
-		str_retreat_weak = "{active} limps away from {inactive}!",
-		str_create = "You press a button on the mobility console labelled 'A'. Through the observation port, you see jointed limbs begin to sprout from the proto-Slimeoid's underside.",
-		str_mobility = "It walks on legs.",
-		str_defeat = "{slimeoid_name}'s knees buckle under it as it collapses to the ground, defeated!",
-		str_walk = "{slimeoid_name} walks along beside you."
-	),
-	EwMobility( # mobility 2
-		id_mobility = "rolling",
-		alias = [
-			"roll",
-			"b"
-		],
-		str_advance = "{active} rolls itself toward {inactive}!",
-		str_retreat = "{active} rolls away from {inactive}!",
-		str_advance_weak = "{active} rolls itself unsteadily towards {inactive}!",
-		str_retreat_weak = "{active} rolls unsteadily away from {inactive}!",
-		str_create = "You press a button on the mobility console labelled 'B'. Through the observation port, you see the proto-Slimeoid becoming smoother and rounder.",
-		str_mobility = "It moves by rolling its body around.",
-		str_defeat = "{slimeoid_name} rolls itself over before collapsing on the ground, defeated!",
-		str_walk = "{slimeoid_name} rolls itself along the ground behind you."
-	),
-	EwMobility( # mobility 3
-		id_mobility = "flagella",
-		alias = [
-			"flagella",
-			"tendrils",
-			"tentacles",
-			"c"
-		],
-		str_advance = "{active} slithers toward {inactive}!",
-		str_retreat = "{active} slithers away from {inactive}!",
-		str_advance_weak = "{active} drags itself toward {inactive}!",
-		str_retreat_weak = "{active} drags itself away from {inactive}!",
-		str_create = "You press a button on the mobility console labelled 'C'. Through the observation port, you see masses of writhing flagella begin to protrude from the proto-Slimeoid's extremities.",
-		str_mobility = "It moves by pulling itself around with its flagella.",
-		str_defeat = "{slimeoid_name}'s flagella go limp as it collapses to the ground, defeated!",
-		str_walk = "{slimeoid_name} writhes its way along the ground on its flagella next to you."
-	),
-	EwMobility( # mobility 4
-		id_mobility = "jets",
-		alias = [
-			"fluid",
-			"jet",
-			"d"
-		],
-		str_advance = "{active} propels itself toward {inactive}!",
-		str_retreat = "{active} propels itself away from {inactive}!",
-		str_advance_weak = "{active} sputters towards {inactive}!",
-		str_retreat_weak = "{active} sputters away from {inactive}!",
-		str_create = "You press a button on the mobility console labelled 'D'. Through the observation port, you see openings forming in the proto-Slimeoid's body, which begin to vent fluid.",
-		str_mobility = "It moves via jet-propulsion by squirting fluids.",
-		str_defeat = "{slimeoid_name} fires its fluid jets wildly in a panic until it completely deflates and collapses, defeated!",
-		str_walk = "{slimeoid_name} tries to keep pace with you, spurting jets of fluid to propel itself along behind you."
-	),
-	EwMobility( # mobility 5
-		id_mobility = "slug",
-		alias = [
-			"undulate",
-			"e"
-		],
-		str_advance = "{active} undulates its way toward {inactive}!",
-		str_retreat = "{active} undulates itself away from {inactive}!",
-		str_advance_weak = "{active} heaves itself slowly toward {inactive}!",
-		str_retreat_weak = "{active} heaves itself slowly away from {inactive}!",
-		str_create = "You press a button on the mobility console labelled 'E'. Through the observation port, you see the base of the proto-Slimeoid's body widen and flatten out.",
-		str_mobility = "It moves like a slug, undulating its underside along the ground.",
-		str_defeat = "{slimeoid_name} stops moving entirely and collapses to the ground, defeated!",
-		str_walk = "{slimeoid_name} glacially drags its way along behind you in its slug-like way. Your walk ends up taking fucking forever."
-	),
-	EwMobility( # mobility 5
-		id_mobility = "float",
-		alias = [
-			"gas",
-			"f"
-		],
-		str_advance = "{active} floats toward {inactive}!",
-		str_retreat = "{active} floats away from {inactive}!",
-		str_advance_weak = "{active} bobs unsteadily through the air towards {inactive}!",
-		str_retreat_weak = "{active} bobs unsteadily away from {inactive}!",
-		str_create = "You press a button on the mobility console labelled 'F'. Through the observation port, you see the of the proto-Slimeoid's body start to inflate itself.",
-		str_mobility = "It floats in the air with the use of super-low-density gas bladders.",
-		str_defeat = "{slimeoid_name} spins wildly in the air before careening to the ground, defeated!",
-		str_walk = "{slimeoid_name} bobs along next to you on its leash like a balloon."
-	),
-	EwMobility( # mobility 5
-		id_mobility = "wings",
-		alias = [
-			"fly",
-			"g"
-		],
-		str_advance = "{active} darts through the air toward {inactive}!",
-		str_retreat = "{active} flaps away from {inactive}!",
-		str_advance_weak = "{active} flaps its way doggedly towards {inactive}!",
-		str_retreat_weak = "{active} flaps doggedly away from {inactive}!",
-		str_create = "You press a button on the mobility console labelled 'G'. Through the observation port, you see the proto-slimeoid start to sprout wide, flat, thin appendages.",
-		str_mobility = "It moves by making short flights through the air with its wings.",
-		str_defeat = "{slimeoid_name} flaps helplessly as it spins out and crashes into the ground, defeated!",
-		str_walk = "{slimeoid_name} flaps along through the air next to you, occasionally perching in trees or windowsills along the route."
-	)
-]
-
-# A map of id_mobility to EwBody objects.
-mobility_map = {}
-
-# A list of mobility names
-mobility_names = []
-
-# Populate mobility map, including all aliases.
-for mobility in mobility_list:
-	mobility_map[mobility.id_mobility] = mobility
-	mobility_names.append(mobility.id_mobility)
-
-	for alias in mobility.alias:
-		mobility_map[alias] = mobility
-
-# All offense attributes in the game.
-offense_list = [
-	EwOffense( # offense 1
-		id_offense = slimeoid_weapon_blades,
-		alias = [
-			"edged",
-			"edges",
-			"edgy",
-			"bladed",
-			"blade",
-			"a"
-		],
-		str_attack = "{active} slashes {inactive} with its blades!",
-		str_attack_weak = "{active} desperately swipes at {inactive} with its blades!",
-		str_attack_coup = "{active} slices deep into {inactive}! Green goo splatters onto the ground from the wound!!",
-		str_create = "You press a button on the weapon console labelled 'A'. Through the observation port, you see long, sharp protrusions begin to form on the proto-Slimeoid's extremities.",
-		str_offense = "It slices foes with retractible blades.",
-		str_observe = "{slimeoid_name} is sharpening its retractible blades on a stone nearby."
-	),
-	EwOffense( # offense 2
-		id_offense = slimeoid_weapon_teeth,
-		alias = [
-			"bite",
-			"biting",
-			"crunch",
-			"crunching",
-			"b"
-		],
-		str_attack = "{active} sinks its teeth into {inactive}!",
-		str_attack_weak = "{active} gnashes its teeth, biting {inactive} wherever it can!",
-		str_attack_coup = "{active} bites hard into {inactive}, tearing off a piece and chewing it hungrily!",
-		str_create = "You press a button on the weapon console labelled 'B'. Through the observation port, you see large bony structures resembling teeth forming in the proto-Slimeoid's... mouth?",
-		str_offense = "It can bite foes with deadly fangs.",
-		str_observe = "{slimeoid_name} is idly picking its sharp teeth."
-	),
-	EwOffense( # offense 3
-		id_offense = slimeoid_weapon_grip,
-		alias = [
-			"squeeze",
-			"grab",
-			"squeezing",
-			"grabbing",
-			"gripping",
-			"constrict",
-			"constriction",
-			"c"
-		],
-		str_attack = "{active} grabs {inactive} and squeezes hard!",
-		str_attack_weak = "{active} grabs at {inactive}, trying to fend it off!",
-		str_attack_coup = "{active} grips {inactive} like a vice, squeezing until you hear a sickening pop!",
-		str_create = "You press a button on the weapon console labelled 'C'. Through the observation port, you see the proto-Slimeoid's limbs becoming thicker and stronger, beginning to twist and writhe, seeking something to grip onto.",
-		str_offense = "It can grab and crush its foes with its limbs.",
-		str_observe = "{slimeoid_name} picks up a rock off the ground and squeezes it like a stress ball."
-	),
-	EwOffense( # offense 4
-		id_offense = slimeoid_weapon_bludgeon,
-		alias = [
-			"strike",
-			"striking",
-			"smash",
-			"smashing",
-			"bash",
-			"bashing",
-			"crush",
-			"crushing",
-			"d"
-		],
-		str_attack = "{active} bashes {inactive} with its limbs!",
-		str_attack_weak = "{active} flails its limbs to strike back at {inactive}!",
-		str_attack_coup = "{active} winds back and smashes {inactive}, dealing a knockout blow!",
-		str_create = "You press a button on the weapon console labelled 'D'. Through the observation port, you see the ends of the proto-Slimeoid's limbs becoming harder and heavier.",
-		str_offense = "It can smash foes with one or more of its limbs.",
-		str_observe = "{slimeoid_name} spots an insect on the ground nearby and smashes it."
-	),
-	EwOffense( # offense 5
-		id_offense = slimeoid_weapon_spikes,
-		alias = [
-			"puncture",
-			"spear",
-			"e"
-		],
-		str_attack = "{active} skewers {inactive} with its spikes!",
-		str_attack_weak = "{active} tries to defend itself from {inactive} with its spikes!",
-		str_attack_coup = "{active} punctures {inactive} with its spikes, opening a hole that oozes green fluid all over the ground!",
-		str_create = "You press a button on the weapon console labelled 'E'. Through the observation port, you see hard spikes forming out of the congealing slime biomatter.",
-		str_offense = "It can puncture its enemies with the spikes on its body.",
-		str_observe = "{slimeoid_name} carefully adjusts its position so as not to prick itself with its own spikes."
-	),
-	EwOffense( # offense 6
-		id_offense = slimeoid_weapon_electricity,
-		alias = [
-			"strike",
-			"f"
-		],
-		str_attack = "{active} unleashes a pent-up electrical discharge into {inactive}!",
-		str_attack_weak = "{active} sparks and flickers with electricity, shocking {inactive}!",
-		str_attack_coup = "{active} charges up and sends a bolt of electricity through {inactive}, making it sizzle!",
-		str_create = "You press a button on the weapon console labelled 'F'. Through the observation port, you see the proto-Slimeoid begin to spark with small electrical discharges.",
-		str_offense = "It crackles with stored electrical energy.",
-		str_observe = "A fly flies a little too near {slimeoid_name} and is zapped with a tiny bolt of electricity, killing it instantly."
-	),
-	EwOffense( # offense 7
-		id_offense = slimeoid_weapon_slam,
-		alias = [
-			"bodyslam",
-			"g"
-		],
-		str_attack = "{active} slams its entire body into {inactive}!",
-		str_attack_weak = "{active} flails itself back against {inactive}'s onslaught!",
-		str_attack_coup = "{active} hurls its whole weight into {inactive}, crushing it to the ground!",
-		str_create = "You press a button on the weapon console labelled 'G'. Through the observation port, you see the ends of the proto-Slimeoid's congealing body condense, becoming heavier and more robust.",
-		str_offense = "It can slam its body into its foes with tremendous force.",
-		str_observe = "{slimeoid_name} shifts its weight back and forth before settling down in a kind of sumo-squat position."
-	)
-]
-
-# A map of id_offense to EwBody objects.
-offense_map = {}
-
-# A list of offense names
-offense_names = []
-
-# Populate offense map, including all aliases.
-for offense in offense_list:
-	offense_map[offense.id_offense] = offense
-	offense_names.append(offense.id_offense)
-
-	for alias in offense.alias:
-		offense_map[alias] = offense
-
-
-
-
-# All defense attributes in the game.
-defense_list = [
-	EwDefense( # defense 1
-		id_defense = slimeoid_armor_scales,
-		alias = [
-			"scale",
-			"scaled",
-			"scaly",
-			"a"
-		],
-		str_defense = "",
-		str_pet = "You carefully run your hand over {slimeoid_name}'s hide, making sure to go with the grain so as not to slice your fingers open on its sharp scales.",
-		str_abuse = "You pick up {slimeoid_name} by the hind legs, swinging them over your head and repeatedly slamming them to the ground.",
-		str_create = "You press a button on the armor console labelled 'A'. Through the observation port, you see the proto-Slimeoid's skin begin to glint as it sprouts roughly-edged scales.",
-		str_armor = "It is covered in scales.",
-		id_resistance = slimeoid_weapon_electricity,
-		id_weakness = slimeoid_special_TK,
-		str_resistance = " {}'s scales conduct the electricity away from its vitals!",
-		str_weakness = " {}'s scales refract and amplify the disrupting brainwaves inside its skull!",
-	),
-	EwDefense( # defense 2
-		id_defense = slimeoid_armor_boneplates,
-		alias = [
-			"bone",
-			"bony",
-			"bones",
-			"plate",
-			"plates",
-			"armor",
-			"plating",
-			"b"
-		],
-		str_defense = "",
-		str_pet = "You pat one of the hard, bony plates covering {slimeoid_name}'s skin.",
-		str_abuse = "You take a stick and hit {slimeoid_name}'s face with it.",
-		str_create = "You press a button on the armor console labelled 'B'. Through the observation port, you see hard bony plates begin to congeal on the proto-Slimeoid's surface.",
-		str_armor = "It is covered in bony plates.",
-		id_resistance = slimeoid_weapon_blades,
-		id_weakness = slimeoid_special_spines,
-		str_resistance = " {}'s bone plates block the worst of the damage!",
-		str_weakness = " {}'s bone plates only drive the quills deeper into its body as it moves!",
-
-	),
-	EwDefense( # defense 3
-		id_defense = slimeoid_armor_quantumfield,
-		alias = [
-			"quantum",
-			"field",
-			"energy",
-			"c"
-		],
-		str_defense = "",
-		str_pet = "You pat {slimeoid_name}, and your hand tingles as it passes through the quantum field that surrounds its body.",
-		str_abuse = "You grab hold of {slimeoid_name} and shake them aggressively, until their quantum static is splayed out and they look nauseous.",
-		str_create = "You press a button on the armor console labelled 'C'. Through the observation port, start to notice the proto-Slimeoid begin to flicker, and you hear a strange humming sound.",
-		str_armor = "It is enveloped in a field of quantum uncertainty.",
-		id_resistance = slimeoid_weapon_slam,
-		id_weakness = slimeoid_special_laser,
-		str_resistance = " {}'s quantum superposition makes it difficult to hit head-on!",
-		str_weakness = " {}'s quantum particles are excited by the high-frequency radiation, destabilizing its structure!",
-
-	),
-	EwDefense( # defense 4
-		id_defense = slimeoid_armor_formless,
-		alias = [
-			"amorphous",
-			"shapeless",
-			"squishy",
-			"d"
-		],
-		str_defense = "",
-		str_pet = "You pat {slimeoid_name}, its fluid, shapeless body squishing and deforming in response to even slight pressure.",
-		str_abuse = "You stick your fist into {slimeoid_name}'s squishy body and jostle its vital organs.",
-		str_create = "You press a button on the armor console labelled 'D'. Through the observation port, you see the proto-Slimeoid suddenly begin to twist itself, stretching and contracting as its shape rapidly shifts.",
-		str_armor = "It is malleable and can absorb blows with ease.",
-		id_resistance = slimeoid_weapon_bludgeon,
-		id_weakness = slimeoid_special_webs,
-		str_resistance = " {}'s squishy body easily absorbs the blows!",
-		str_weakness = " {}'s squishy body easily adheres to and becomes entangled by the webs!",
-
-	),
-	EwDefense( # defense 5
-		id_defense = slimeoid_armor_regeneration,
-		alias = [
-			"healing",
-			"regen",
-			"e"
-		],
-		str_defense = "",
-		str_pet = "You pat {slimeoid_name}. Its skin is hot, and you can feel it pulsing rhythmically.",
-		str_abuse = "You take a blowtorch to {slimeoid_name}'s pulsating skin and watch as it wears itself out trying to regenerate the rapidly burning tissue.",
-		str_create = "You press a button on the armor console labelled 'E'. Through the observation port, you see the proto-Slimeoid begin to pulse, almost like a beating heart.",
-		str_armor = "It can regenerate damage to its body rapidly.",
-		id_resistance = slimeoid_weapon_spikes,
-		id_weakness = slimeoid_special_spit,
-		str_resistance = " {} quickly begins regenerating the small puncture wounds inflicted by the spikes!",
-		str_weakness = " {}'s regeneration is impeded by the corrosive chemicals!",
-
-	),
-	EwDefense( # defense 6
-		id_defense = slimeoid_armor_stench,
-		alias = [
-			"stink",
-			"smell",
-			"f"
-		],
-		str_defense = "",
-		str_pet = "You pat {slimeoid_name}, taking care not to inhale through your nose, as one whiff of its odor has been known to make people lose their lunch.",
-		str_abuse = "You try to humiliate {slimeoid_name} by plugging your nose and kicking it around.",
-		str_create = "You press a button on the armor console labelled 'F'. Through the observation port, you see the proto-Slimeoid give off bubbles of foul-colored gas.",
-		str_armor = "It exudes a horrible stench.",
-		id_resistance = slimeoid_weapon_teeth,
-		id_weakness = slimeoid_special_throw,
-		str_resistance = " {}'s noxious fumes make its opponent hesitant to put its mouth anywhere near it!",
-		str_weakness = " {}'s foul odor gives away its position, making it easy to target with thrown projectiles!",
-
-	),
-	EwDefense( # defense 7
-		id_defense = slimeoid_armor_oil,
-		alias = [
-			"slick",
-			"g"
-		],
-		str_defense = "",
-		str_pet = "You pat {slimeoid_name}'s slick wet skin, and your hand comes away coated in a viscous, slippery oil.",
-		str_abuse = "You turn {slimeoid_name} upside down and push their oily body, sending them slipping down the sidewalk.",
-		str_create = "You press a button on the armor console labelled 'G'. Through the observation port, you see the surface of the proto-Slimeoid become shiny with some kind of oily fluid.",
-		str_armor = "It is covered in a coating of slippery oil.",
-		id_resistance = slimeoid_weapon_grip,
-		id_weakness = slimeoid_special_fire,
-		str_resistance = " {}'s slippery coating makes it extremely difficult to grab on to!",
-		str_weakness = " {}'s oily coating is flammable, igniting as it contacts the flame!",
-
-	)
-]
-
-# A map of id_defense to EwBody objects.
-defense_map = {}
-
-# A list of defense names
-defense_names = []
-
-# Populate defense map, including all aliases.
-for defense in defense_list:
-	defense_map[defense.id_defense] = defense
-	defense_names.append(defense.id_defense)
-
-	for alias in defense.alias:
-		defense_map[alias] = defense
-
-# All special attributes in the game.
-special_list = [
-	EwSpecial( # special 1
-		id_special = slimeoid_special_spit,
-		alias = [
-			"spitting",
-			"spray",
-			"squirt",
-			"spraying",
-			"squirting",
-			"liquid",
-			"fluid",
-			"acid",
-			"acidic",
-			"toxic",
-			"poison",
-			"a"
-		],
-		str_special_attack = "{active} spits acidic ooze onto {inactive}!",
-		str_special_attack_weak = "{active} coughs and spurts up a sputtering spray of acid at {inactive}!",
-		str_special_attack_coup = "{active} vomits a torrent of acid onto {inactive}, deteriorating it to the point that it can no longer fight!",
-		str_create = "You press a button on the special attack console labelled 'A'. Through the observation port, you see the proto-Slimeoid's body begin to excrete a foul, toxic ooze.",
-		str_special = "It can spit acidic ooze.",
-		str_observe = "A bit of acidic fluid drips from {slimeoid_name} onto the ground, where it smokes and sizzles."
-	),
-	EwSpecial( # special 2
-		id_special = slimeoid_special_laser,
-		alias = [
-			"beam",
-			"energy",
-			"radiation",
-			"b"
-		],
-		str_special_attack = "{active} sears {inactive} with a blast of radiation!",
-		str_special_attack_weak = "{active} starts to flicker before firing an unsteady beam of light at {inactive}!",
-		str_special_attack_coup = "{active} blasts {inactive} with a beam of green energy, searing it all over its body!",
-		str_create = "You press a button on the special attack console labelled 'B'. Through the observation port, you see the proto-Slimeoid's body begin to glow with energy as the gestation vat's built-in Geiger Counter begins to click frantically.",
-		str_special = "It can fire beams of radiation.",
-		str_observe = "{slimeoid_name} suddenly glows with radioactive energy. Best not to look directly at it until it settles down..."
-	),
-	EwSpecial( # special 3
-		id_special = slimeoid_special_spines,
-		alias = [
-			"spikes",
-			"spiky",
-			"spiny",
-			"quills",
-			"c"
-		],
-		str_special_attack = "{active} fires a volley of quills into {inactive}!",
-		str_special_attack_weak = "{active} desperately fires a few of its last quills into {inactive}!",
-		str_special_attack_coup = "{active} fires a rapid burst of sharp quills into {inactive}, filling it like a pincushion!",
-		str_create = "You press a button on the special attack console labelled 'C'. Through the observation port, you see the proto-Slimeoid's congealing body suddenly protruding with long, pointed spines, which quickly retract back into it.",
-		str_special = "It can fire sharp quills.",
-		str_observe = "{slimeoid_name} shudders and ejects a few old quills onto the ground. You can see new ones already growing in to replace them."
-	),
-	EwSpecial( # special 4
-		id_special = slimeoid_special_throw,
-		alias = [
-			"throwing",
-			"hurling",
-			"hurl",
-			"d"
-		],
-		str_special_attack = "{active} picks up a nearby {object} and hurls it into {inactive}!",
-		str_special_attack_weak = "{active} unsteadily hefts a nearby {object} before throwing it into {inactive}!",
-		str_special_attack_coup = "{active} hurls a {object}, which smashes square into {inactive}, knocking it to the ground! A direct hit!",
-		str_create = "You press a button on the special attack console labelled 'D'. Through the observation port, you see the proto-Slimeoid's limbs become more articulate.",
-		str_special = "It can hurl objects at foes.",
-		str_observe = "{slimeoid_name} is idly picking up stones and seeing how far it can toss them."
-	),
-	EwSpecial( # special 5
-		id_special = slimeoid_special_TK,
-		alias = [
-			"telekinesis",
-			"psychic",
-			"e"
-		],
-		str_special_attack = "{active} focuses on {inactive}... {inactive} convulses in pain!",
-		str_special_attack_weak = "{active}'s cranium bulges and throbs! {inactive} convulses!",
-		str_special_attack_coup = "{active} emanates a strange static sound as {inactive} is inexplicably rendered completely unconscious!",
-		str_create = "You press a button on the special attack console labelled 'E'. You momentarily experience an uncomfortable sensation, sort of like the feeling you get when you know there's a TV on in the room even though you can't see it.",
-		str_special = "It can generate harmful frequencies with its brainwaves.",
-		str_observe = "You momentarily black out. When you come to, your nose is bleeding. {slimeoid_name} tries to look innocent."
-	),
-	EwSpecial( # special 6
-		id_special = slimeoid_special_fire,
-		alias = [
-			"chemical",
-			"breath",
-			"breathe",
-			"f"
-		],
-		str_special_attack = "{active} ejects a stream of fluid which ignites in the air, burning {inactive}!",
-		str_special_attack_weak = "{active} fires an unsteady, sputtering stream of fluid that ignites and singes {inactive}!",
-		str_special_attack_coup = "{active} empties its fluid bladders in a final burst of liquid! {inactive} is completely engulfed in the conflagration!",
-		str_create = "You press a button on the special attack console labelled 'F'. Through the observation port, you see fluid bladders forming deep under the still-forming proto-Slimeoid's translucent skin.",
-		str_special = "It can fire a stream of pyrophoric fluid at its foes.",
-		str_observe = "A bit of fluid drips from {slimeoid_name} onto the floor and ignites, but you manage to smother the small flame quickly before it spreads."
-	),
-	EwSpecial( # special 7
-		id_special = slimeoid_special_webs,
-		alias = [
-			"webbing",
-			"web",
-			"g"
-		],
-		str_special_attack = "{active} fires a stream of sticky webbing onto {inactive}!",
-		str_special_attack_weak = "{active} is running out of webbing! It shoots as much as it can onto {inactive}!",
-		str_special_attack_coup = "{active} gathers itself up before spurting a blast of webbing that coats {inactive}'s body, completely ensnaring it!",
-		str_create = "You press a button on the special attack console labelled 'G'. Through the observation port, you see large glands forming near the surface of the still-forming proto-Slimeoid's translucent skin.",
-		str_special = "It can spin webs and shoot webbing fluid to capture prey.",
-		str_observe = "{slimeoid_name} is over in the corner, building itself a web to catch prey in."
-	)
-]
-
-# A map of id_special to EwBody objects.
-special_map = {}
-
-# A list of special names
-special_names = []
-
-# Populate special map, including all aliases.
-for special in special_list:
-	special_map[special.id_special] = special
-	special_names.append(special.id_special)
-
-	for alias in special.alias:
-		special_map[alias] = special
-
-def get_strat_a(combat_data, in_range, first_turn, active):
-	base_attack = 30
-	base_evade = 10
-	base_block = 10
-
-	weight_attack = base_attack
-	weight_evade = base_evade
-	weight_block = base_block
-
-	if in_range:
-		if active:
-			weight_attack *= 3
-			weight_block *= 2
-		else:
-			weight_block *= 3
-
-	else:
-		if active:
-			weight_evade *= 2
-		else:
-			weight_evade *= 5
-
-	strat = random.randrange(weight_attack + weight_evade + weight_block)
-	if strat < weight_attack:
-		strat_used = slimeoid_strat_attack
-	elif strat < weight_attack + weight_evade:
-		strat_used = slimeoid_strat_evade
-	else:
-		strat_used = slimeoid_strat_block
-
-	if first_turn:
-		sap_spend = int(random.triangular(0, combat_data.sap, int(combat_data.sap * 0.8))) + 1
-
-	else:
-		sap_spend = combat_data.sap
-
-	sap_spend = min(sap_spend, combat_data.sap)
-
-	return strat_used, sap_spend
-
-def get_strat_b(combat_data, in_range, first_turn, active):
-	base_attack = 20
-	base_evade = 10
-	base_block = 10
-
-	weight_attack = base_attack
-	weight_evade = base_evade
-	weight_block = base_block
-
-	if in_range:
-		if active:
-			weight_attack *= 3
-			weight_block *= 2
-		else:
-			weight_block *= 2
-			weight_evade *= 3
-
-	else:
-		if active:
-			weight_attack *= 3
-			weight_evade *= 3
-		else:
-			weight_evade *= 4
-			weight_block *= 2
-
-	strat = random.randrange(weight_attack + weight_evade + weight_block)
-	if strat < weight_attack:
-		strat_used = slimeoid_strat_attack
-	elif strat < weight_attack + weight_evade:
-		strat_used = slimeoid_strat_evade
-	else:
-		strat_used = slimeoid_strat_block
-
-	if first_turn:
-		sap_spend = int(random.triangular(0, combat_data.sap, int(combat_data.sap * 0.65))) + 1
-
-	else:
-		sap_spend = combat_data.sap
-
-	sap_spend = min(sap_spend, combat_data.sap)
-
-	return strat_used, sap_spend
-
-def get_strat_c(combat_data, in_range, first_turn, active):
-	base_attack = 30
-	base_evade = 10
-	base_block = 10
-
-	weight_attack = base_attack
-	weight_evade = base_evade
-	weight_block = base_block
-
-	if in_range:
-		if active:
-			weight_attack *= 3
-		else:
-			weight_block *= 2
-			weight_evade *= 2
-
-	else:
-		if active:
-			weight_attack *= 3
-		else:
-			weight_evade *= 2
-			weight_block *= 2
-
-	strat = random.randrange(weight_attack + weight_evade + weight_block)
-	if strat < weight_attack:
-		strat_used = slimeoid_strat_attack
-	elif strat < weight_attack + weight_evade:
-		strat_used = slimeoid_strat_evade
-	else:
-		strat_used = slimeoid_strat_block
-
-	if first_turn:
-		sap_spend = int(random.triangular(0, combat_data.sap, int(combat_data.sap * 0.8))) + 1
-
-	else:
-		sap_spend = combat_data.sap
-
-	sap_spend = min(sap_spend, combat_data.sap)
-
-	return strat_used, sap_spend
-
-def get_strat_d(combat_data, in_range, first_turn, active):
-	base_attack = 20
-	base_evade = 5
-	base_block = 15
-
-	weight_attack = base_attack
-	weight_evade = base_evade
-	weight_block = base_block
-
-	if in_range:
-		if active:
-			weight_attack *= 3
-		else:
-			weight_attack /= 2
-			weight_block *= 2
-
-	else:
-		if active:
-			weight_attack *= 3
-		else:
-			weight_attack /= 2
-			weight_block *= 2
-
-	strat = random.randrange(weight_attack + weight_evade + weight_block)
-	if strat < weight_attack:
-		strat_used = slimeoid_strat_attack
-	elif strat < weight_attack + weight_evade:
-		strat_used = slimeoid_strat_evade
-	else:
-		strat_used = slimeoid_strat_block
-
-	if first_turn:
-		sap_spend = int(random.triangular(0, combat_data.sap, int(combat_data.sap * 0.35))) + 1
-
-	else:
-		sap_spend = combat_data.sap
-
-	sap_spend = min(sap_spend, combat_data.sap)
-
-	return strat_used, sap_spend
-
-def get_strat_e(combat_data, in_range, first_turn, active):
-	base_attack = 20
-	base_evade = 10
-	base_block = 10
-
-	weight_attack = base_attack
-	weight_evade = base_evade
-	weight_block = base_block
-
-	if in_range:
-		if active:
-			weight_attack *= 2
-			weight_evade *= 2
-		else:
-			weight_evade *= 4
-
-	else:
-		if active:
-			weight_attack *= 4
-			weight_block *= 2
-		else:
-			weight_block *= 3
-
-	strat = random.randrange(weight_attack + weight_evade + weight_block)
-	if strat < weight_attack:
-		strat_used = slimeoid_strat_attack
-	elif strat < weight_attack + weight_evade:
-		strat_used = slimeoid_strat_evade
-	else:
-		strat_used = slimeoid_strat_block
-
-	if first_turn:
-		sap_spend = int(random.triangular(0, combat_data.sap, int(combat_data.sap * 0.65))) + 1
-
-	else:
-		sap_spend = combat_data.sap
-
-	sap_spend = min(sap_spend, combat_data.sap)
-
-	return strat_used, sap_spend
-
-def get_strat_f(combat_data, in_range, first_turn, active):
-	base_attack = 20
-	base_evade = 20
-	base_block = 10
-
-	weight_attack = base_attack
-	weight_evade = base_evade
-	weight_block = base_block
-
-	if in_range:
-		if active:
-			weight_attack *= 3
-			weight_evade *= 2
-		else:
-			weight_evade *= 3
-			weight_block *= 2
-
-	else:
-		if active:
-			weight_attack *= 4
-			weight_block *= 2
-		else:
-			weight_block *= 3
-			weight_evade *= 2
-
-
-	strat = random.randrange(weight_attack + weight_evade + weight_block)
-	if strat < weight_attack:
-		strat_used = slimeoid_strat_attack
-	elif strat < weight_attack + weight_evade:
-		strat_used = slimeoid_strat_evade
-	else:
-		strat_used = slimeoid_strat_block
-
-	if first_turn:
-		sap_spend = int(random.triangular(0, combat_data.sap, int(combat_data.sap * 0.35))) + 1
-
-	else:
-		sap_spend = combat_data.sap
-
-	sap_spend = min(sap_spend, combat_data.sap)
-
-	return strat_used, sap_spend
-
-def get_strat_g(combat_data, in_range, first_turn, active):
-	base_attack = 10
-	base_evade = 15
-	base_block = 5
-
-	weight_attack = base_attack
-	weight_evade = base_evade
-	weight_block = base_block
-
-	if in_range:
-		if active:
-			weight_attack *= 4
-		else:
-			weight_evade *= 2
-
-	else:
-		if active:
-			weight_attack *= 4
-		else:
-			weight_evade *= 2
-
-	strat = random.randrange(weight_attack + weight_evade + weight_block)
-	if strat < weight_attack:
-		strat_used = slimeoid_strat_attack
-	elif strat < weight_attack + weight_evade:
-		strat_used = slimeoid_strat_evade
-	else:
-		strat_used = slimeoid_strat_block
-
-	if first_turn:
-		sap_spend = int(random.triangular(0, combat_data.sap, int(combat_data.sap * 0.2))) + 1
-
-	else:
-		sap_spend = combat_data.sap
-
-	sap_spend = min(sap_spend, combat_data.sap)
-
-	return strat_used, sap_spend
-# All brain attributes in the game.
-brain_list = [
-	EwBrain( # brain 1
-		id_brain = "a",
-		alias = [
-			"typea",
-			"type a"
-		],
-		str_create = "You press a button on the brain console labelled 'A'.\nA cluster of neurons coagulates within your newly-forming Slimeoid. The proto-Slimeoid begins to move, thrashing about as if in frustration.",
-		str_brain = "It is extremely irritable.",
-		str_observe = "{slimeoid_name} is snarling. You're not sure if it's angry at you, or at the world in general.",
-		str_pet = "{slimeoid_name} hisses at you.",
-		str_walk = "You wrestle {slimeoid_name} down and force a leash onto it, as it angrily snarls and hisses at you in protest.",
-		str_feed = "{slimeoid_name} almost bites your hand off as you offer the {food_name} to it! It growls at you before eating, as if to secure its prey.",
-		str_kill = "{slimeoid_name} howls with savage delight at the bloodshed!!",
-		str_death = "{slimeoid_name} howls in fury at its master's death! It tears away in a blind rage!",
-		str_victory = "{slimeoid_name} roars in triumph!!",
-		str_battlecry = "{slimeoid_name} roars with bloodlust!! ",
-		str_battlecry_weak = "{slimeoid_name} is too breathless to roar, but is still filled with bloodlust!! ",
-		str_movecry = "{slimeoid_name} snarls at its prey! " ,
-		str_movecry_weak = "{slimeoid_name}  hisses with frustrated rage! ",
-		str_revive = "{slimeoid_name} howls at your return, annoyed to have been kept waiting.",
-		str_spawn = "{slimeoid_name} shakes itself off to get rid of some excess gestation fluid, then starts to hiss at you. Seems like a real firecracker, this one.",
-		str_dissolve = "{slimeoid_name} hisses and spits with fury as you hold it over the N.L.A.C.U Dissolution Vats. Come on, get in there...\n{slimeoid_name} claws at you, clutching at the edge of the vat, screeching with rage even as you hold its head under the surface and wait for the chemical soup to do its work. At last, it stops fighting.\n\n{slimeoid_name} is no more.",
-		get_strat = get_strat_a,
-		str_abuse= "{slimeoid_name} lashes out, trying to fight back."
-
-	),
-	EwBrain( # brain 2
-		id_brain = "b",
-		alias = [
-			"typeb",
-			"type b"
-		],
-		str_create = "You press a button on the brain console labelled 'B'.\nA cluster of neurons coagulates within your newly-forming Slimeoid. The proto-Slimeoid begins to move about its gestation tank, exploring its surroundings.",
-		str_brain = "It is enthusiastic about almost everything.",
-		str_observe = "{slimeoid_name} notices you looking at it and seems delighted!",
-		str_pet = "{slimeoid_name} purrs happily.",
-		str_walk = "{slimeoid_name} is so excited for its walk, it can barely hold still enough to let you put the leash on it!",
-		str_feed = "{slimeoid_name} starts running circles around you and drooling uncontrollably in anticipation as soon as you reach for the {food_name}.",
-		str_kill = "{slimeoid_name} gives a bestial woop of excitement for your victory!",
-		str_death = "{slimeoid_name} gives a wail of grief at its master's death, streaking away from the scene.",
-		str_victory = "{slimeoid_name} woops with delight at its victory!",
-		str_battlecry = "{slimeoid_name} lets out a loud war woop! " ,
-		str_battlecry_weak = "{slimeoid_name} is determined not to lose! ",
-		str_movecry = "{slimeoid_name} is thrilled by the battle! " ,
-		str_movecry_weak = "{slimeoid_name} seems a little less thrilled now... ",
-		str_revive = "{slimeoid_name} is waiting patiently downtown when you return from your time as a corpse. It knew you'd be back!",
-		str_spawn = "{slimeoid_name} gets up off the ground slowly at first, but then it notices you and leaps into your arms. It sure seems glad to see you!",
-		str_dissolve = "You order {slimeoid_name} into the Dissolution Vats. It's initially confused, but realization of what you're asking slowly crawks across its features.\nIt doesn't want to go, but after enough stern commanding, it finally pitches itself into the toxic sludge, seemingly too heartbroken to fear death.\n\n{slimeoid_name} is no more.",
-		get_strat = get_strat_b,
-		str_abuse = "{slimeoid_name} yelps in pain. It's beside itself and doesn't know what to do..."
-	),
-	EwBrain( # brain 3
-		id_brain = "c",
-		alias = [
-			"typec",
-			"type c"
-		],
-		str_create = "You press a button on the brain console labelled 'C'.\nA cluster of neurons coagulates within your newly-forming Slimeoid. The proto-Slimeoid draws its congealing body together, as if trying to gather its strength.",
-		str_brain = "It is quiet and withdrawn.",
-		str_observe = "{slimeoid_name} seems to be resting, possibly deep in thought.",
-		str_pet = "{slimeoid_name} doesn't react.",
-		str_walk = "{slimeoid_name} holds still as you place the leash on it. It regards the leash, seemingly pontificating.",
-		str_feed = "{slimeoid_name} shows neither happiness nor reluctance as you offer the {food_name}. It accepts the treat as though it were a mere formality.",
-		str_kill = "{slimeoid_name} regards the corpse of your former adversary with an unknowable expression.",
-		str_death = "{slimeoid_name} stares at the killer, memorizing their face before fleeing the scene.",
-		str_victory = "{slimeoid_name} silently turns away from its defeated opponent.",
-		str_battlecry = "{slimeoid_name} carefully regards its opponent. ",
-		str_battlecry_weak = "{slimeoid_name} tries to steady itself. ",
-		str_movecry = "{slimeoid_name} seems to be getting impatient. " ,
-		str_movecry_weak = "{slimeoid_name} is losing its composure just a little! ",
-		str_revive = "{slimeoid_name} is downtown when you return from the sewers. You find it staring silently up at ENDLESS WAR.",
-		str_spawn = "{slimeoid_name} regards you silently from the floor. You can't tell if it likes you or not, but it starts to follow you regardless.",
-		str_dissolve = "You pick up {slimeoid_name} and hurl it into the N.L.A.C.U. Dissolution Vats before it starts to suspect anything. It slowly sinks into the chemical soup, kind of like Arnold at the end of Terminator 2, only instead of giving you a thumbs-up, it stares at you with an unreadable expression. Betrayal? Confusion? Hatred? Yeah, probably.\n\n{slimeoid_name} is no more.",
-		get_strat = get_strat_c,
-		str_abuse = "You can feel {slimeoid_name} trembling, but it just takes your abuse."
-	),
-	EwBrain( # brain 4
-		id_brain = "d",
-		alias = [
-			"typed",
-			"type d"
-		],
-		str_create = "You press a button on the brain console labelled 'D'.\nA cluster of neurons coagulates within your newly-forming Slimeoid. The proto-Slimeoid lazily turns over in its gestation vat, floating and doing little else.",
-		str_brain = "It is usually staring off into space.",
-		str_observe = "{slimeoid_name} stares off into the distance. Who knows if it's actually looking at anything in particular.",
-		str_pet = "{slimeoid_name} is startled out of a stupor by your touch.",
-		str_walk = "{slimeoid_name} hardly seems to notice you fastening it with a leash.",
-		str_feed = "You have to literally shove the {food_name} into {slimeoid_name}'s face to get its attention. It takes a moment to recover its orientation before accepting the treat.",
-		str_kill = "{slimeoid_name} wasn't paying attention and missed the action.",
-		str_death = "{slimeoid_name} is startled to realize its master has died. It blinks in confusion before fleeing.",
-		str_victory = "{slimeoid_name} keeps attacking for a moment before realizing it's already won.",
-		str_battlecry = "{slimeoid_name} is weighing its options! ",
-		str_battlecry_weak = "{slimeoid_name} is desperately trying to come up with a plan! ",
-		str_movecry = "{slimeoid_name} Isn't really feeling this. " ,
-		str_movecry_weak = "{slimeoid_name} tries to buy itself some time to think! ",
-		str_revive = "{slimeoid_name} is exactly where you left it when you died.",
-		str_spawn = "{slimeoid_name} flops over on the floor and stares up at you. Its gaze wanders around the room for a while before it finally picks itself up to follow you.",
-		str_dissolve = "You lead {slimeoid_name} up to the edge of the Dissolution Vats and give a quick 'Hey, look, a distraction!'. {slimeoid_name} is immediately distracted and you shove it over the edge. Landing in the vat with a sickening *gloop* sound, it sinks quickly under the fluid surface, flailing madly in confusion and desperation.\n\n{slimeoid_name} is no more.",
-		get_strat = get_strat_d,
-		str_abuse = "You have a hard time with it, {slimeoid_name} is panicked and keeps trying to run."
-	),
-	EwBrain( # brain 5
-		id_brain = "e",
-		alias = [
-			"typee",
-			"type e"
-		],
-		str_create = "You press a button on the brain console labelled 'E'.\nA cluster of neurons coagulates within your newly-forming Slimeoid. The proto-Slimeoid starts to sporadically twitch and shiver.",
-		str_brain = "It is extremely skittish and jumpy.",
-		str_observe = "{slimeoid_name} is glancing around furtively, seemingly scanning for threats.",
-		str_pet = "{slimeoid_name} flinches nervously at your touch.",
-		str_walk = "{slimeoid_name} shivers in place as you're fastening the leash, then starts to pull and tug at it, seemingly perturbed.",
-		str_feed = "{slimeoid_name} backs up anxiously as you reach out with the {food_name} in your hand. You sigh and take a bite of the treat yourself to convince {slimeoid_name} that its not poisoned. It reluctantly accepts the {food_name} and starts nibbling at it.",
-		str_kill = "{slimeoid_name} peers out from behind its master, hoping the violence is over.",
-		str_death = "{slimeoid_name} is overcome with terror, skittering away from the killer in a mad panic!",
-		str_victory = "{slimeoid_name} is deeply relieved that the battle is over.",
-		str_battlecry = "{slimeoid_name} chitters fearfully! ",
-		str_battlecry_weak = "{slimeoid_name} squeals in abject terror! ",
-		str_movecry = "{slimeoid_name} makes a break for it! " ,
-		str_movecry_weak = "{slimeoid_name} is in a full-blown panic! ",
-		str_revive = "{slimeoid_name} peeks out from behind some trash cans before rejoining you. It seems relieved to have you back.",
-		str_spawn = "{slimeoid_name}'s eyes dart frantically around the room. Seeing you, it darts behind you, as if for cover from an unknown threat.",
-		str_dissolve = "{slimeoid_name} is looking around the lab nervously, obviously unnerved by the Slimeoid technology. Its preoccupation makes it all too easy to lead it to the Dissolution Vats and kick its legs out from under it, knocking it in. As it falls and hits the solvent chemicals, it wails and screeches in shock and terror, but the noise eventually quiets as it dissolves into a soft lump, then disintegrates altogether.\n\n{slimeoid_name} is no more.",
-		get_strat = get_strat_e,
-		str_abuse = "{slimeoid_name} is squealing in horror. It will remember this..."
-	),
-	EwBrain( # brain 6
-		id_brain = "f",
-		alias = [
-			"typef",
-			"type f"
-		],
-		str_create = "You press a button on the brain console labelled 'F'.\nA cluster of neurons coagulates within your newly-forming Slimeoid. The proto-Slimeoid darts to the opposite side of the gestation vat. You're not sure, but you feel like it's watching you.",
-		str_brain = "It acts secretive, as though it's up to something.",
-		str_observe = "{slimeoid_name} is moving around, apparently searching for... something.",
-		str_pet = "{slimeoid_name} seems nonplussed, but doesn't object.",
-		str_walk = "{slimeoid_name} exasperatedly lets you fit it with a leash for a walk.",
-		str_feed = "{slimeoid_name} only seems to halfway pay attention as you offer the {food_name}. It pockets the treat for later and eats it when it thinks you aren't looking.",
-		str_kill = "{slimeoid_name} rifles through your victim's pockets for food.",
-		str_death = "{slimeoid_name} rifles through its dead master's pockets for whatever it can find before slinking away.",
-		str_victory = "{slimeoid_name} shakes itself off after the battle.",
-		str_battlecry = "{slimeoid_name} makes its move! ",
-		str_battlecry_weak = "{slimeoid_name}, backed into a corner, tries to counterattack! ",
-		str_movecry = "{slimeoid_name} decides on a tactical repositioning. " ,
-		str_movecry_weak = "{slimeoid_name} thinks it'd better try something else, and fast! ",
-		str_revive = "{slimeoid_name} starts following you around again not long after you have returned from the dead.",
-		str_spawn = "{slimeoid_name} picks itself up off the floor and regards you coolly. It seems as if it's gauging your usefulness.",
-		str_dissolve = "{slimeoid_name} eyes you suspiciously as you approach the Dissolution Vats. It's on to you. Before it has a chance to bolt, you grab it, hoist it up over your head, and hurl it into the chemical soup. {slimeoid_name} screeches in protest, sputtering and hissing as it thrashes around in the vat, but the chemicals work quickly and it soon dissolves into nothing.\n\n{slimeoid_name} is no more.",
-		get_strat = get_strat_f,
-		str_abuse = "{slimeoid_name}'s fear grows as it desperately looks for an escape."
-	),
-	EwBrain( # brain 7
-		id_brain = "g",
-		alias = [
-			"typeg",
-			"type g"
-		],
-		str_create = "You press a button on the brain console labelled 'G'.\nA cluster of neurons coagulates within your newly-forming Slimeoid. The proto-Slimeoid begins to flit around the gestation vat, seemingly unsure where to go.",
-		str_brain = "It seems to have no idea what it's doing.",
-		str_observe = "{slimeoid_name} seems unsure of whether it wants to wander around or just stay put.",
-		str_pet = "{slimeoid_name} seems confused about how to react.",
-		str_walk = "{slimeoid_name} lets you put its leash on it, but immediately starts to trip over it and get tangled in it.",
-		str_feed = "{slimeoid_name} stares at the {food_name} like it's unfamiliar with the concept of food. You make a chewing motion with your mouth to demonstrate. It still seems confused. You lose your patience and force-feed the treat to your slimeoid.",
-		str_kill = "{slimeoid_name} seems unsure of whether to celebrate the victory or to mourn the decline of your civilization into rampant youth violence.",
-		str_death = "{slimeoid_name} starts to approach its master's body, then changes its mind and starts to run away. It trips over itself and falls on its way out.",
-		str_victory = "{slimeoid_name} looks around, apparently shocked that it somehow won.",
-		str_battlecry = "{slimeoid_name} decides to actually do something for once! ",
-		str_battlecry_weak = "{slimeoid_name} decides to actually do something for once, now that it's probably too late.",
-		str_movecry = "{slimeoid_name} is moving around aimlessly! " ,
-		str_movecry_weak = "{slimeoid_name} is limping around aimlessly! ",
-		str_revive = "{slimeoid_name} wanders by, seemingly by accident, but thinks it probably ought to start following you again.",
-		str_spawn = "{slimeoid_name} starts to pick itself up off the floor, then changes its mind and lies back down. Then it gets up again. Lies down again. Up. Down. Up. Ok, this time it stays up.",
-		str_dissolve = "{slimeoid_name} is perplexed by the laboratory machinery. Taking advantage of its confusion, you point it towards the Dissolution Vats, and it gormlessly meanders up the ramp and over the edge. You hear a gloopy SPLOOSH sound, then nothing. You approach the vats and peer over the edge, but see no trace of your former companion.\n\n{slimeoid_name} is no more.",
-		get_strat = get_strat_g,
-		str_abuse = "{slimeoid_name} still hasn't caught up to all this chaos, and is lost and confused."
-	)
-]
-
-# A map of id_brain to EwBrain objects.
-brain_map = {}
-
-# A list of brain names
-brain_names = []
-
-# Populate brain map, including all aliases.
-for brain in brain_list:
-	brain_map[brain.id_brain] = brain
-	brain_names.append(brain.id_brain)
-
-	for alias in brain.alias:
-		brain_map[alias] = brain
 
 hue_analogous = -1
 hue_neutral = 0
